### 2022-10-11 #4844 @editorialbot
[PRE REVIEW]: FreqAI: generalizing adaptive modeling for chaotic time-series market forecasts
**Submitting author:** @robcaulk (<a href="http://orcid.org/0000-0001-5618-8629">Robert Caulk</a>)
**Repository:** https://github.com/freqtrade/freqtrade
**Branch with paper.md** (empty if default branch): JOSS-paper-submission
**Version:** v1.0
