### 2022-10-12 ad7b78e
Update exchange init to use .pro if available
