### 2022-10-12 96e6c1b
Docs: add `ignore_buying_expired_candle_after` and `order pricing` to summary
