### 2022-10-12 e985c18
Implement basic ccxt.pro to test
