### 2022-10-13 39c27cf
Don't fail if fetchTickers is not availlable
