### 2022-10-13 75f1a12
Move "tickers_needed" check to pairlistmanager to cover all pairlists
