### 2022-10-13 2045780
Reinstate default of 1000% for roi
closes #7583
