### 2022-10-13 #4844 @kthyng
Hi @robcaulk — JOSS submissions are required to be related to research to be in scope. Can you make an argument for how this submission is research software?
