### 2022-10-13 7672586
Fix unreliable hyperopt test
