### 2022-10-13 #4844 @robcaulk
Hello @kthyng,

I am glad you asked. FreqAI is a significant component of Freqtrade, which is a non scientific software, and therefore I can understand your immediate confusion.

However, there is no doubt that FreqAI is a novel scientific software, built to enable advanced hypothesis testing on *live* and *historical* time-series data. In addition to novel components, FreqAI promotes reproducible research, something which is lacking in the field of machine learning. Maybe I can highlight how this research tool can/is being used:

1.) The broadest impact is quite clear:

FreqAI connects open-sourced machine learning libraries. This is one of the most valuable research components - scientists can now test hypotheses 1:1 Catboost vs XGboost vs LightGBM vs torch vs tensorflow (or any other python library desired) with ease on both live and historical data sets. That flexibility serves to strengthen each of the individual libraries and it serves to enable advanced scientific hypothesis testing.

2.) On a more granular level, the FreqAI sand-box provides well-tested frameworks for research in:

* Data set engineering: https://www.freqtrade.io/en/stable/freqai-feature-engineering/#feature-engineering
* Outlier detection: https://www.freqtrade.io/en/stable/freqai-feature-engineering/#outlier-detection
* Advanced data collection, data pre-processing, and data post-processing methods: e.g. https://www.freqtrade.io/en/stable/freqai-feature-engineering/#identifying-outliers-with-the-dissimilarity-index-di

These robust features (currently used by hundreds of people on a daily basis) open up a streamlined and reproducible pipeline of live and historical data for scientists. This is a principal problem in the field of machine learning and time-series modeling - studies are often difficult to benchmark and reproduce due to poorly maintained/tested code. FreqAI changes that by serving as a foundation for hypothesis testing on historical and live data sets. In addition - FreqAI's numerous users have helped to identify bugs and inspire improvements that would certainly be lacking from typical machine learning article supplementary material code.

3.) Novel features. Last, but certainly not least. FreqAI introduces a wide range of novel components that are currently not available or published elsewhere:

* Novel metrics: https://www.freqtrade.io/en/stable/freqai-feature-engineering/#inlier-metric
* Novel algorithmic approach to adaptive modeling: https://www.freqtrade.io/en/stable/freqai/#general-approach
* Novel architecture enabling a massive sand-box: https://www.freqtrade.io/en/stable/freqai-developers/#project-architecture

4.) All of the above is sitting on top of a robust documentation, continuous integration with good testing coverage, and a deep searchable interactive knowledge base (discord). This is a powerful demonstration of open-source machine learning. This community driven project is singular in a field filled with "proprietary models" and "proprietary algorithms."

The arguments here compliment our arguments detailed in our JOSS paper submission, where we outline details of the role that FreqAI plays in open-source machine learning as well as the algorithmic and scientific novelties.

Thanks for your effort, and we look forward to the review,

Robert et al.
