### 2022-10-14 a364b51
add strategy v4
