+++ user_data/strategies/FreqaiBinaryClassStrategy_v4.py
@@ -1,0 +1,553 @@
+from typing import Dict, List, Optional, Tuple, Union
+import logging
+from functools import reduce
+from turtle import update
+from h11 import Data
+from datetime import datetime, timedelta, timezone
+import pandas as pd
+import talib.abstract as ta
+from pandas_ta.trend import adx
+from pandas import DataFrame
+from technical import qtpylib
+import numpy as np
+from scipy.signal import argrelextrema
+from sklearn.metrics import precision_recall_curve
+from freqtrade.exchange import timeframe_to_prev_date
+from freqtrade.persistence import Trade
+from technical.util import resample_to_interval, resampled_merge
+from freqtrade.strategy import DecimalParameter, IntParameter, merge_informative_pair
+from freqtrade.strategy.interface import IStrategy
+
+
+logger = logging.getLogger(__name__)
+
+
+def find_support_levels(df: DataFrame) -> DataFrame:
+    """
+    cond1 = df['Low'][i] < df['Low'][i-1]   
+    cond2 = df['Low'][i] < df['Low'][i+1]   
+    cond3 = df['Low'][i+1] < df['Low'][i+2]   
+    cond4 = df['Low'][i-1] < df['Low'][i-2]  
+    """
+    cond1 = df["low"] < df["low"].shift(1)
+    cond2 = df["low"] < df["low"].shift(-1)
+    cond3 = df["low"].shift(-1) < df["low"].shift(-2)
+    cond4 = df["low"].shift(1) < df["low"].shift(2)
+    return (cond1 & cond2 & cond3 & cond4)
+
+
+def get_max_labels(df: DataFrame, alpha: float = 0.5) -> DataFrame:
+
+    price = (df['high'] + df['low'] + df['close']) / 3
+
+    max_peaks = argrelextrema(price.values, np.greater, order=12)[0]
+
+    out = adx(df["high"], df["low"], df["close"], window=12)
+    diplus = out["DMP_14"]
+
+    di_thr = diplus[max_peaks].mean() + diplus[max_peaks].std() * alpha
+
+    nn = 2
+    labels = np.zeros(len(df), dtype=np.int32)
+    for mp in max_peaks:
+        ref_close = price.iloc[mp]
+        start = max(0, mp-nn)
+        end = min(df.shape[0], mp+nn+1)
+        pct = np.abs(price[start:end] / ref_close - 1)
+        is_close = np.where(pct <= 0.005)[0]
+        left_idx = is_close[0]
+        right_idx = is_close[-1]
+        # locality labeling
+        if diplus[mp-nn+left_idx:mp-nn+right_idx].mean() >= di_thr:
+            labels[mp-nn+left_idx:mp-nn+right_idx] = 1
+    if labels.max() == 0:  # if not any positive label is found, we force it
+        idx = np.nanargmax(diplus[max_peaks])
+        labels[max_peaks[idx]] = 1
+    return labels
+
+
+def get_min_labels(df: DataFrame, alpha : float = 0.5) -> DataFrame:
+
+    price = (df['high'] + df['low'] + df['close']) / 3
+
+    min_peaks = argrelextrema(price.values, np.less, order=12)[0]
+
+    out = adx(df["high"], df["low"], df["close"], window=12)
+    diminus = out["DMN_14"]
+    di_thr = diminus[min_peaks].mean() + diminus[min_peaks].std() * alpha
+    nn = 2
+    labels = np.zeros(len(df), dtype=np.int32)
+    for mp in min_peaks:
+        ref_close = price.iloc[mp]
+        start = max(0, mp-nn)
+        end = min(df.shape[0], mp+nn+1)
+        pct = np.abs(price[start:end] / ref_close - 1)
[…]
