### 2022-10-14 1d8d360
update _search_all_objects functioning
