### 2022-10-14 9d4ba76
Update usages of search_all_objects
