+++ freqtrade/resolvers/strategy_resolver.py
@@ -268,6 +268,14 @@ class StrategyResolver(IResolver):
             "or contains Python code errors."
         )
 
+    @classmethod
+    def build_search_paths(cls, config: Config, user_subdir: Optional[str] = None,
+                           extra_dirs: List[str] = []) -> List[Path]:
+
+        if 'strategy_path' in config:
+            extra_dirs = [config['strategy_path']] + extra_dirs
+        return super().build_search_paths(config, user_subdir, extra_dirs)
+
 
 def warn_deprecated_setting(strategy: IStrategy, old: str, new: str, error=False):
     if hasattr(strategy, old):
