### 2022-10-14 4bfe587
Generalize "path" variables for resolvers
