### 2022-10-15 4982897
Fix catboost tests polluting CWD
