### 2022-10-15 05ca725
Remove no longer needed local state
