### 2022-10-15 b6c096d
Simplify backtest condition
