### 2022-10-15 99dbba6
avoid reading from disk to instantiate large objects
