### 2022-10-16 6cb1414
Fix random test failure due to catboost bug
https://github.com/catboost/catboost/issues/2195
