### 2022-10-16 b7dce8d
Update stoploss.md
