### 2022-10-16 e82baf5
Use helper-method to modify trades_open
