+++ freqtrade/optimize/backtesting.py
@@ -1125,7 +1125,7 @@ class Backtesting:
                         open_trade_count -= 1
                         open_trade_count_start -= 1
                         open_trades[pair].remove(t)
-                        LocalTrade.trades_open.remove(t)
+                        LocalTrade.remove_bt_trade(t)
                         self.wallets.update()
 
                 # 2. Process entries.
