### 2022-10-16 0e8cf36
Keep trade state in LocalTrade
