+++ tests/test_persistence.py
@@ -2406,6 +2406,8 @@ def test_Trade_object_idem():
         'get_trading_volume',
 
     )
+    EXCLUDES2 = ('trades', 'trades_open', 'bt_trades_open_pp', 'bt_open_open_trade_count',
+                 'total_profit')
 
     # Parent (LocalTrade) should have the same attributes
     for item in trade:
@@ -2416,7 +2418,7 @@ def test_Trade_object_idem():
     # Fails if only a column is added without corresponding parent field
     for item in localtrade:
         if (not item.startswith('__')
-                and item not in ('trades', 'trades_open', 'total_profit')
+                and item not in EXCLUDES2
                 and type(getattr(LocalTrade, item)) not in (property, FunctionType)):
             assert item in trade
 
