### 2022-10-16 82d75d8
Fix tests that failed due to new strategy.
