### 2022-10-16 8534dfb
Extract backtest 1 candle from main function
