### 2022-10-16 6252ae4
Convert position_stacking to attribute of backtest
