### 2022-10-17 600b886
Bump lightgbm from 3.3.2 to 3.3.3
Bumps [lightgbm](https://github.com/microsoft/LightGBM) from 3.3.2 to 3.3.3.
- [Release notes](https://github.com/microsoft/LightGBM/releases)
- [Commits](https://github.com/microsoft/LightGBM/compare/v3.3.2...v3.3.3)

---
updated-dependencies:
- dependency-name: lightgbm
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
