### 2022-10-17 7ec1e3b
Bump fastapi from 0.85.0 to 0.85.1
Bumps [fastapi](https://github.com/tiangolo/fastapi) from 0.85.0 to 0.85.1.
- [Release notes](https://github.com/tiangolo/fastapi/releases)
- [Commits](https://github.com/tiangolo/fastapi/compare/0.85.0...0.85.1)

---
updated-dependencies:
- dependency-name: fastapi
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
