### 2022-10-17 abe4d32
Update wording in get_analyzed_dataframe docs
