### 2022-10-17 943f5f2
Improve get_pair_dataframe doc wording
