### 2022-10-17 c2914fe
Don't fail contract size repopulation if pair is no longer available
