### 2022-10-17 c8e103e
Adjust typehints to match return value
