### 2022-10-17 #4844 @robcaulk
Hey @Fei-Tao, thanks for taking the time to oversee this review of our JOSS submission. I reviewed the list that you linked to and found some good reviewers:

bnriiitb
mkareshk
jameschapman19
makyol
edrictam
zhenmuw
daviddewhurst
ady00
Annielytix
max-little
shagunsodhani
simonpcouch

Please do not hesitate to let me know if you need me to extend this list (I found quite a few suitable candidates).

Cheers,

Robert et al.
