### 2022-10-18 c3d4fb9
Simplify backtest calling interface
