### 2022-10-18 033c5bd
Make check constant pred labels agnostic
