### 2022-10-18 d427226
Update docker_quickstart.md
