### 2022-10-18 34ccada
Tests with seperate thread
