+++ user_data/strategies/bb_scalp_15m.py
@@ -179,7 +179,7 @@ class bb_scalp_15m(WAOStrategy):
                 # & (dataframe['ema_9'] > dataframe['sma_200'])
                 # & (dataframe['macdhist'] > -0.08)
             ),
-            'entry'] = 1
+            'buy'] = 1
 
         return dataframe
 
@@ -196,5 +196,5 @@ class bb_scalp_15m(WAOStrategy):
                 # | (qtpylib.crossed_below(dataframe['ema_9'], dataframe['sma_200']))
 
             ),
-            'exit'] = 1
+            'sell'] = 1
         return dataframe
