### 2022-10-19 6e95b66
Modify test ensuring we always have a 0 index
