### 2022-10-20 295ba21
Make extra_returns_per_train values available during backtest
