### 2022-10-20 073ce16
remove un-used f-string
