### 2022-10-20 a9db668
avoid redundant indicator population for corr_pair list
