### 2022-10-20 5899440
fix typos - live predictions candles
