### 2022-10-20 6606a01
refactoring - remove unnecessary config file
