+++ freqtrade/freqai/data_kitchen.py
@@ -1,7 +1,7 @@
 import copy
 import logging
 import shutil
-from datetime import datetime, timezone
+from datetime import datetime, timedelta, timezone
 from math import cos, sin
 from pathlib import Path
 from typing import Any, Dict, List, Tuple
@@ -21,7 +21,6 @@ from freqtrade.configuration import TimeRange
 from freqtrade.constants import Config
 from freqtrade.exceptions import OperationalException
 from freqtrade.exchange import timeframe_to_seconds
-from freqtrade.freqai import freqai_util
 from freqtrade.strategy.interface import IStrategy
 
 
@@ -84,16 +83,17 @@ class FreqaiDataKitchen:
         self.backtest_live_models = config.get("freqai_backtest_live_models", False)
 
         if not self.live:
-            self.full_path = freqai_util.get_full_models_path(self.config)
-            self.full_timerange = self.create_fulltimerange(
-                self.config["timerange"], self.freqai_config.get("train_period_days", 0)
-            )
+            self.full_path = self.get_full_models_path(self.config)
 
             if self.backtest_live_models:
-                self.set_timerange_from_ready_models()
-                (self.training_timeranges,
-                 self.backtesting_timeranges) = self.split_timerange_live_models()
+                if self.pair:
+                    self.set_timerange_from_ready_models()
+                    (self.training_timeranges,
+                     self.backtesting_timeranges) = self.split_timerange_live_models()
             else:
+                self.full_timerange = self.create_fulltimerange(
+                    self.config["timerange"], self.freqai_config.get("train_period_days", 0)
+                )
                 (self.training_timeranges, self.backtesting_timeranges) = self.split_timerange(
                     self.full_timerange,
                     config["freqai"]["train_period_days"],
@@ -117,7 +117,7 @@ class FreqaiDataKitchen:
         :param metadata: dict = strategy furnished pair metadata
         :param trained_timestamp: int = timestamp of most recent training
         """
-        self.full_path = freqai_util.get_full_models_path(self.config)
+        self.full_path = self.get_full_models_path(self.config)
         self.data_path = Path(
             self.full_path
             / f"sub-train-{pair.split('/')[0]}_{trained_timestamp}"
@@ -1300,10 +1300,102 @@ class FreqaiDataKitchen:
     def set_timerange_from_ready_models(self):
         backtesting_timerange, \
             assets_end_dates = (
-                freqai_util.get_timerange_and_assets_end_dates_from_ready_models(self.full_path))
+                self.get_timerange_and_assets_end_dates_from_ready_models(self.full_path))
 
         self.backtest_live_models_data = {
             "backtesting_timerange": backtesting_timerange,
             "assets_end_dates": assets_end_dates
             }
         return
+
+    def get_full_models_path(self, config: Config) -> Path:
+        """
+        Returns default FreqAI model path
+        :param config: Configuration dictionary
+        """
+        freqai_config: Dict[str, Any] = config["freqai"]
+        return Path(
+            config["user_data_dir"] / "models" / str(freqai_config.get("identifier"))
+        )
+
+    def get_timerange_and_assets_end_dates_from_ready_models(
+            self, models_path: Path) -> Tuple[TimeRange, Dict[str, Any]]:
+        """
+        Returns timerange information based on a FreqAI model directory
+        :param models_path: FreqAI model path
+
+        :return: a Tuple with (Timerange calculated from directory and
+        a Dict with pair and model end training dates info)
+        """
+        all_models_end_dates = []
+        assets_end_dates: Dict[str, Any] = self.get_assets_timestamps_training_from_ready_models(
+                                           models_path)
+        for key in assets_end_dates:
+            for model_end_date in assets_end_dates[key]:
[…]
