### 2022-10-20 107845a
Keep version number in docker versioning
