### 2022-10-20 f4814a7
Improve test resiliance to small roundings
