### 2022-10-20 bd424a8
Add Trade from_json method
