### 2022-10-21 d159188
add missing datetime conversion in fromJson
