### 2022-10-22 5ee3b8c
update config recording to use all configs, fix tests
