### 2022-10-22 0ff7a07
Move price_jump_warn to dataloading
it's not relevant for live data, and should only run when loading data
from disk.
