### 2022-10-22 547fd28
Price-jump detection should only run once
