### 2022-10-22 84a194b
Simplify stoploss logic by removing redundant condition
