### 2022-10-22 c4a2ee0
fix freqai test
