### 2022-10-23 3d7a311
removed sleep calls, better channel sending
