+++ freqtrade/rpc/api_server/ws/proxy.py
@@ -16,6 +16,10 @@ class WebSocketProxy:
         self._websocket: Union[FastAPIWebSocket, WebSocket] = websocket
 
     @property
+    def raw(self):
+        return self._websocket
+
+    @property
     def remote_addr(self) -> Tuple[Any, ...]:
         if isinstance(self._websocket, WebSocket):
             return self._websocket.remote_address
