+++ wao/wao_strategy_controller.py
@@ -1,8 +1,10 @@
-from wao.brain_util import perform_execute_buy, perform_execute_sell, write_to_backtest_table, clear_cumulative_value, create_initial_account_balance_binance_file, is_romeo_alive, remove_from_pool
+import sys
+sys.path.append("wao/")
+from brain_util import perform_execute_buy, perform_execute_sell, write_to_backtest_table, clear_cumulative_value, create_initial_account_balance_binance_file, is_romeo_alive, remove_from_pool
 import threading
-from wao.brain_config import BrainConfig
-from wao.brain_util import create_watchers
-from wao.notifier import send_start_deliminator_message
+from brain_config import BrainConfig
+from brain_util import create_watchers
+from notifier import send_start_deliminator_message
 
 from execution.config import Config
 
