+++ wao/_429_watcher.py
@@ -4,7 +4,9 @@ import watchdog.observers
 import os
 from pathlib import Path
 
-from wao.notifier import post_request
+import sys
+sys.path.append("wao/")
+from notifier import post_request
 
 
 class _429_Watcher(watchdog.events.PatternMatchingEventHandler):
