### 2022-10-23 3468edd
Add enable_ws config setting
