### 2022-10-23 34c0874
allow custom exit for V4
