### 2022-10-23 c361415
Simplify "refresh" condition
