### 2022-10-23 d057146
Improve test for worker throttle
