+++ tests/test_worker.py
@@ -1,6 +1,9 @@
 import logging
 import time
+from datetime import timedelta
 from unittest.mock import MagicMock, PropertyMock
+
+import time_machine
 
 from freqtrade.data.dataprovider import DataProvider
 from freqtrade.enums import State
@@ -59,11 +62,46 @@ def test_throttle(mocker, default_conf, caplog) -> None:
     end = time.time()
 
     assert result == 42
-    assert end - start > 0.1
+    assert 0.3 > end - start > 0.1
     assert log_has_re(r"Throttling with 'throttled_func\(\)': sleep for \d\.\d{2} s.*", caplog)
 
     result = worker._throttle(throttled_func, throttle_secs=-1)
     assert result == 42
+
+
+def test_throttle_sleep_time(mocker, default_conf, caplog) -> None:
+
+    caplog.set_level(logging.DEBUG)
+    worker = get_patched_worker(mocker, default_conf)
+    sleep_mock = mocker.patch("freqtrade.worker.Worker._sleep")
+    with time_machine.travel("2022-09-01 05:00:00 +00:00") as t:
+        def throttled_func(x=1):
+            t.shift(timedelta(seconds=x))
+            return 42
+
+        assert worker._throttle(throttled_func, throttle_secs=5) == 42
+        # This moves the clock by 1 second
+        assert sleep_mock.call_count == 1
+        assert 3.8 < sleep_mock.call_args[0][0] < 4.1
+
+        sleep_mock.reset_mock()
+        # This moves the clock by 1 second
+        assert worker._throttle(throttled_func, throttle_secs=10) == 42
+        assert sleep_mock.call_count == 1
+        assert 8.8 < sleep_mock.call_args[0][0] < 9.1
+
+        sleep_mock.reset_mock()
+        # This moves the clock by 5 second, so we only throttle by 5s
+        assert worker._throttle(throttled_func, throttle_secs=10, x=5) == 42
+        assert sleep_mock.call_count == 1
+        assert 4.8 < sleep_mock.call_args[0][0] < 5.1
+
+        t.move_to("2022-09-01 05:01:00 +00:00")
+        sleep_mock.reset_mock()
+        # Throttle for more than 5m (1 timeframe)
+        assert worker._throttle(throttled_func, throttle_secs=400, x=5) == 42
+        assert sleep_mock.call_count == 1
+        assert 394.8 < sleep_mock.call_args[0][0] < 395.1
 
 
 def test_throttle_with_assets(mocker, default_conf) -> None:
