### 2022-10-23 f93b6ee
Improve timing for worker throttling
