### 2022-10-23 94b65a0
fix message typing in channel manager, minor improvements
