### 2022-10-23 07e813d
change param record to only get certain params
