### 2022-10-23 4d2b7a7
move record params to utils, use rapidjson
