+++ wao/notifier.py
@@ -1,6 +1,5 @@
 import sys
-sys.path.append("wao/")
-from brain_config import BrainConfig
+from wao.brain_config import BrainConfig
 import requests
 
 from execution.config import Config
