+++ wao/brain_util.py
@@ -3,10 +3,9 @@ import watchdog
 import os
 import time
 import sys
-sys.path.append("wao/")
-from brain_config import BrainConfig
-from _429_watcher import _429_Watcher
-from error_watcher import Error_Watcher
+from wao.brain_config import BrainConfig
+from wao._429_watcher import _429_Watcher
+from wao.error_watcher import Error_Watcher
 import pickle
 
 from execution.config import Config
