+++ wao/wao_strategy_controller_test.py
@@ -1,6 +1,5 @@
 import sys
-sys.path.append("wao/")
-from wao_strategy_controller import WAOStrategyController
+from wao.wao_strategy_controller import WAOStrategyController
 import time
 import datetime
 brain = 'Freq_bb_15m'
