+++ wao/_429_watcher.py
@@ -5,8 +5,7 @@ import os
 from pathlib import Path
 
 import sys
-sys.path.append("wao/")
-from notifier import post_request
+from wao.notifier import post_request
 
 
 class _429_Watcher(watchdog.events.PatternMatchingEventHandler):
