+++ wao/error_watcher.py
@@ -45,7 +45,6 @@ def smooth_romeo_restart(error_line):
         romeo.perform_sell_signal(RomeoExitPriceType.SS)
         romeo.send_error_report(error_line)
     else:
-        post_request(error_line)
         send_to_trello(title=error_line,description=error_line)
 
 
@@ -89,6 +88,8 @@ def get_error_line(file_name):
 def send_to_trello(title,description):
     notifier = Notifier(BrainConfig.MODE)
     notifier.create_trello_bug_ticket(title,description)
+    notifier.post_request(description,is_from_error_report=True)
+
 
 class Error_Watcher(watchdog.events.PatternMatchingEventHandler):
 
