+++ wao/_429_watcher.py
@@ -4,7 +4,6 @@ import watchdog.observers
 import os
 from pathlib import Path
 
-import sys
 from wao.notifier import post_request
 
 
