+++ wao/wao_strategy_controller.py
@@ -1,4 +1,3 @@
-import sys
 from wao.brain_util import perform_execute_buy, perform_execute_sell, write_to_backtest_table, clear_cumulative_value, create_initial_account_balance_binance_file, is_romeo_alive, remove_from_pool
 import threading
 from wao.brain_config import BrainConfig
