### 2022-10-24 96f4de4
Bump progressbar2 from 4.0.0 to 4.1.1
Bumps [progressbar2](https://github.com/WoLpH/python-progressbar) from 4.0.0 to 4.1.1.
- [Release notes](https://github.com/WoLpH/python-progressbar/releases)
- [Changelog](https://github.com/wolph/python-progressbar/blob/develop/CHANGES.rst)
- [Commits](https://github.com/WoLpH/python-progressbar/compare/v4.0.0...v4.1.1)

---
updated-dependencies:
- dependency-name: progressbar2
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
