+++ requirements-hyperopt.txt
@@ -6,4 +6,4 @@ scipy==1.9.2
 scikit-learn==1.1.2
 scikit-optimize==0.9.0
 filelock==3.8.0
-progressbar2==4.0.0
+progressbar2==4.1.1
