### 2022-10-24 54d029d
Bump ccxt from 2.0.25 to 2.0.58
Bumps [ccxt](https://github.com/ccxt/ccxt) from 2.0.25 to 2.0.58.
- [Release notes](https://github.com/ccxt/ccxt/releases)
- [Changelog](https://github.com/ccxt/ccxt/blob/master/exchanges.cfg)
- [Commits](https://github.com/ccxt/ccxt/compare/2.0.25...2.0.58)

---
updated-dependencies:
- dependency-name: ccxt
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
