### 2022-10-24 0328cd5
Bump scipy from 1.9.2 to 1.9.3
Bumps [scipy](https://github.com/scipy/scipy) from 1.9.2 to 1.9.3.
- [Release notes](https://github.com/scipy/scipy/releases)
- [Commits](https://github.com/scipy/scipy/compare/v1.9.2...v1.9.3)

---
updated-dependencies:
- dependency-name: scipy
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
