+++ requirements-hyperopt.txt
@@ -2,7 +2,7 @@
 -r requirements.txt
 
 # Required for hyperopt
-scipy==1.9.2
+scipy==1.9.3
 scikit-learn==1.1.2
 scikit-optimize==0.9.0
 filelock==3.8.0
