### 2022-10-24 51be455
remove np object, make default str
