+++ wao/_429_watcher.py
@@ -4,11 +4,7 @@ import watchdog.observers
 import os
 from pathlib import Path
 
-# Revert
-import sys
-sys.path.append("wao/")
-from notifier import post_request
-# Revert
+from wao.notifier import post_request
 
 
 class _429_Watcher(watchdog.events.PatternMatchingEventHandler):
