### 2022-10-25 f70c00d
Improve variance around worker timing test
