+++ requirements.txt
@@ -38,7 +38,7 @@ sdnotify==0.3.2
 
 # API Server
 fastapi==0.85.1
-pydantic>=1.8.0
+pydantic==1.10.2
 uvicorn==0.18.3
 pyjwt==2.6.0
 aiofiles==22.1.0
