+++ wao/error_watcher.py
@@ -40,10 +40,12 @@ def smooth_romeo_restart(error_line):
     error_line = "[REPORT TO TRELLO]" + error_line
     error_line += (" [SENDING SS]" if is_romeo_alive else " [POOL EMPTY. NO ROMEO FOUND]")
     post_request(error_line, is_from_error_handler=True)
-    romeo.send_error_report(error_line)
 
     if is_romeo_alive:
         romeo.perform_sell_signal(RomeoExitPriceType.SS)
+        romeo.send_error_report(error_line) #todo
+    else:
+        post_request(error_line) #todo uncomment
 
 def string_to_list(string):
     return list(string.split("\n"))
