+++ wao/error_watcher.py
@@ -43,9 +43,9 @@ def smooth_romeo_restart(error_line):
 
     if is_romeo_alive:
         romeo.perform_sell_signal(RomeoExitPriceType.SS)
-        romeo.send_error_report(error_line) #todo
+        romeo.send_error_report(error_line)
     else:
-        post_request(error_line) #todo uncomment
+        post_request(error_line)
 
 def string_to_list(string):
     return list(string.split("\n"))
