### 2022-10-25 3fa5007
Don't use pydantic to type-verify outgoing messages
