+++ wao/error_watcher.py
@@ -39,7 +39,6 @@ def smooth_romeo_restart(error_line):
     is_romeo_alive = romeo is not None
     error_line = "[REPORT TO TRELLO]" + error_line
     error_line += (" [SENDING SS]" if is_romeo_alive else " [POOL EMPTY. NO ROMEO FOUND]")
-    post_request(error_line, )
 
     if is_romeo_alive:
         romeo.perform_sell_signal(RomeoExitPriceType.SS)
