### 2022-10-25 d94c003
Add missing import to hyperopt docs
