### 2022-10-25 604f966
Enhance documentation with full space override sample
closes #7643
