### 2022-10-25 217add7
add strat and config for testing on PR
