### 2022-10-25 b9bf9ed
update rapidjson opts
