### 2022-10-26 cf6b75a
Use combined stoploss_adjust where possible
