### 2022-10-26 6e0ca05
Update function-head for _get_stop_params
