### 2022-10-26 255f385
Simplify stoploss behavior by combining more commonalities
