### 2022-10-26 1c98640
Delete MultiTargetClassifierTestStrategy.py
