+++ user_data/strategies/MultiTargetClassifierTestStrategy.py
@@ -1,244 +1,0 @@
-import logging
-from functools import reduce
-
-import numpy as np
-import pandas as pd
-import talib.abstract as ta
-from pandas import DataFrame
-from technical import qtpylib
-
-from freqtrade.strategy import CategoricalParameter, IStrategy, merge_informative_pair
-
-
-logger = logging.getLogger(__name__)
-
-
-class MultiTargetClassifierTestStrategy(IStrategy):
-    """
-    Example strategy showing how the user connects their own
-    IFreqaiModel to the strategy. Namely, the user uses:
-    self.freqai.start(dataframe, metadata)
-
-    to make predictions on their data. populate_any_indicators() automatically
-    generates the variety of features indicated by the user in the
-    canonical freqtrade configuration file under config['freqai'].
-    """
-
-    minimal_roi = {"0": 0.1, "240": -1}
-
-    plot_config = {
-        "main_plot": {},
-        "subplots": {
-            "prediction": {"prediction": {"color": "blue"}},
-            "do_predict": {
-                "do_predict": {"color": "brown"},
-            },
-        },
-    }
-
-    process_only_new_candles = True
-    stoploss = -0.05
-    use_exit_signal = True
-    # this is the maximum period fed to talib (timeframe independent)
-    startup_candle_count: int = 40
-    can_short = False
-
-    std_dev_multiplier_buy = CategoricalParameter(
-        [0.75, 1, 1.25, 1.5, 1.75], default=1.25, space="buy", optimize=True)
-    std_dev_multiplier_sell = CategoricalParameter(
-        [0.75, 1, 1.25, 1.5, 1.75], space="sell", default=1.25, optimize=True)
-
-    def populate_any_indicators(
-        self, pair, df, tf, informative=None, set_generalized_indicators=False
-    ):
-        """
-        Function designed to automatically generate, name and merge features
-        from user indicated timeframes in the configuration file. User controls the indicators
-        passed to the training/prediction by prepending indicators with `'%-' + coin `
-        (see convention below). I.e. user should not prepend any supporting metrics
-        (e.g. bb_lowerband below) with % unless they explicitly want to pass that metric to the
-        model.
-        :param pair: pair to be used as informative
-        :param df: strategy dataframe which will receive merges from informatives
-        :param tf: timeframe of the dataframe which will modify the feature names
-        :param informative: the dataframe associated with the informative pair
-        """
-
-        coin = pair.split('/')[0]
-
-        if informative is None:
-            informative = self.dp.get_pair_dataframe(pair, tf)
-
-        # first loop is automatically duplicating indicators for time periods
-        for t in self.freqai_info["feature_parameters"]["indicator_periods_candles"]:
-
-            t = int(t)
-            informative[f"%-{coin}rsi-period_{t}"] = ta.RSI(informative, timeperiod=t)
-            informative[f"%-{coin}mfi-period_{t}"] = ta.MFI(informative, timeperiod=t)
-            informative[f"%-{coin}adx-period_{t}"] = ta.ADX(informative, timeperiod=t)
-            informative[f"%-{coin}sma-period_{t}"] = ta.SMA(informative, timeperiod=t)
-            informative[f"%-{coin}ema-period_{t}"] = ta.EMA(informative, timeperiod=t)
-
-            bollinger = qtpylib.bollinger_bands(
-                qtpylib.typical_price(informative), window=t, stds=2.2
-            )
-            informative[f"{coin}bb_lowerband-period_{t}"] = bollinger["lower"]
-            informative[f"{coin}bb_middleband-period_{t}"] = bollinger["mid"]
-            informative[f"{coin}bb_upperband-period_{t}"] = bollinger["upper"]
-
[…]
