### 2022-10-26 5e533b5
Emit a simple "new candle" message to allow UI's to refresh charts
