+++ freqtrade/exchange/exchange.py
@@ -1998,7 +1998,7 @@ class Exchange:
 
         return (
             (self._pairs_last_refresh_time.get((pair, timeframe, candle_type), 0)
-             + interval_in_sec) <= arrow.utcnow().int_timestamp
+             + interval_in_sec) < arrow.utcnow().int_timestamp
         )
 
     @retrier_async
