### 2022-10-27 a541943
fix ood do_predict issue;
