### 2022-10-28 e2b5671
remove double log
