+++ freqtrade/__init__.py
@@ -1,5 +1,5 @@
 """ Freqtrade bot """
-__version__ = '2022.10.dev'
+__version__ = '2022.11.dev'
 
 if 'dev' in __version__:
     try:
