### 2022-10-29 c23a947
Move exchange utilities into separate module
