### 2022-10-29 352adaf
Improve readability of is_time_to_refresh function
