### 2022-10-29 5946678
Remove loop of normalization from metadata
