+++ wao/brain_util.py
@@ -25,9 +25,6 @@ def perform_execute_buy(coin, brain, time_out_hours, dup):
     elif BrainConfig.MODE == Config.MODE_PROD:
         is_test_mode = False
 
-    Config.IS_SCHEDULE_ORDER = BrainConfig.IS_SCHEDULE_ORDER
-    Config.IS_LIMIT_STOP_ORDER_ENABLED = BrainConfig.IS_LIMIT_STOP_ORDER_ENABLED
-    Config.WORKSPACE_NORMAL = BrainConfig.WORKSPACE_NORMAL
     Config.COIN = coin
     Config.BRAIN = brain
     Config.ROMEO_SS_TIMEOUT_HOURS = time_out_hours
