### 2022-10-29 650bb8b
ensure full pair string is used for caching dataframes. If not, revert to old behavior. Update docs.
