### 2022-10-30 39a8947
leverage list length knowledge, f-string change
