### 2022-10-30 a323acf
Improve ShufflePairlist to shuffle only once per candle
