### 2022-10-30 7053f81
simplified predict and predict_proba using super(). Added duplicate class label check.
