### 2022-10-30 bd051cb
Remove pandas workaround for pi image
