+++ requirements.txt
@@ -1,7 +1,5 @@
 numpy==1.23.4
-pandas==1.5.1; platform_machine != 'armv7l'
-# Piwheels doesn't have 1.5.0 yet.
-pandas==1.4.3; platform_machine == 'armv7l'
+pandas==1.5.1
 pandas-ta==0.3.14b
 
 ccxt==2.0.58
