### 2022-10-30 d59a7fa
remove analysis_lock and realign example hybrid strat
