### 2022-10-30 a49edfb
add tests for CatboostClassifier
