### 2022-10-30 #4864 @shagunsodhani
Hey @Fei-Tao, I could use some advice from you. The submitted package, [FreqAI](https://github.com/freqtrade/freqtrade/tree/develop/freqtrade/freqai) is a part of a larger project, [FreqTrade](https://github.com/freqtrade/freqtrade). It is a little tricky to ascertain the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html?highlight=substantial#substantial-scholarly-effort). Overall, I am pretty confident that the work passes this criterion. Still, I wanted to double-check if there is anything else I should keep in mind when reviewing this criterion.
