### 2022-10-30 freqtrade/freqtrade#7668 @shagunsodhani
Questions related to JOSS Review
Hi there!

Thank you for submitting your library to JOSS. I am one of the reviewers and had some follow up questions for the items in the [checklist](https://github.com/openjournals/joss-reviews/issues/4864#issuecomment-1296239688).

1. Regarding "Contribution and authorship: Has the submitting author (@robcaulk) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?", it is clear that you are one of the major contributors. I am surprised that some of the other contributors (Juha Nykänen, Johan van der Vlugt and Richárd Józsa) did not appear as authors on the paper. Two of these people did end up in the acknowledgements. Given that one person was left, I wanted to double check if this might be an oversight?

2. Regarding "A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?", I noticed that the paper covers this aspect clearly but it is not so obvious on the website / github. Maybe it will help to expand on this in the website / github as well.

Some general comments for improving the accompanying paper:

1. Line 25, "the data is freely available", could the authors add some references / links to get access to such data?

2. Line 55 "novel methods which are not available anywhere else..." This sound like a very strong claim. Could the authors provide more details around the unique features (going deeper into the bullet points).

3. Line 182 "publication in scientific journal..." Could the authors add more details / links here?

4. Line 193 "FreqAI has no official sponsers", are there some unofficial sponsers? Wouldn't Github sponsorships count ?
