+++ wao/error_watcher.py
@@ -3,6 +3,7 @@ import threading
 
 import watchdog.events
 import watchdog.observers
+import time
 
 from wao.notifier import post_request
 from wao.brain_config import BrainConfig
@@ -68,10 +69,13 @@ def check_condition(file_name):
 def __check_condition(file_name):
     error_line = get_error_line(file_name)
     if error_line is not None and not is_freqtrade_error(error_line) and not is_timed_out_error(error_line):
-        if BrainConfig.IS_SMOOTH_ERROR_HANDLING_ENABLED:
+        is_throttle_hit = time.time() - BrainConfig.PREVIOUS_ERROR_TIMESTAMP_SECONDS > 3
+        if BrainConfig.IS_SMOOTH_ERROR_HANDLING_ENABLED and is_throttle_hit:
             smooth_romeo_restart(error_line)
         else:
             stop_bot(error_line)
+            send_to_trello(title= "[STOPBOT] "+error_line, description= "is_throttle_hit=" + str(is_throttle_hit) + " " + error_line)
+        BrainConfig.PREVIOUS_ERROR_TIMESTAMP_SECONDS = time.time()
 
 
 def populate_last_error_lines(line_lower):
@@ -88,7 +92,8 @@ def get_error_line(file_name):
         for line in list_of_lines:
             line_str = str(line)
             line_lower = line_str.lower()
-            if ("error" in line_lower or "exception" in line_lower) and (line_lower not in BrainConfig.LAST_ERROR_LINES):
+            if ("error" in line_lower or "exception" in line_lower) and (
+                    line_lower not in BrainConfig.LAST_ERROR_LINES):
                 populate_last_error_lines(line_lower)
                 return line_str
     return None
@@ -98,6 +103,11 @@ def send_to_trello_and_telegram(title, description):
     notifier = Notifier(BrainConfig.MODE)
     notifier.create_trello_bug_ticket(title, description)
     notifier.post_request(description, is_from_error_report=True)
+
+
+def send_to_trello(title, description):
+    notifier = Notifier(BrainConfig.MODE)
+    notifier.create_trello_bug_ticket(title, description)
 
 
 class Error_Watcher(watchdog.events.PatternMatchingEventHandler):
