+++ wao/brain_config.py
@@ -25,3 +25,4 @@ class BrainConfig:
     INITIAL_ACCOUNT_BALANCE_BINANCE_FILE_PATH = FREQTRADE_PATH + "/_initial_account_balance_binance.txt"
     IS_SMOOTH_ERROR_HANDLING_ENABLED = True
     LAST_ERROR_LINES = []
+    PREVIOUS_ERROR_TIMESTAMP_SECONDS = 0
