### 2022-10-31 b3f612e
Bump ccxt to 2.0.96
