### 2022-10-31 freqtrade/freqtrade#7668 @robcaulk
Hey @shagunsodhani, thanks for taking the time to review our software for JOSS.

> I am surprised that some of the other contributors (Juha Nykänen, Johan van der Vlugt and Richárd Józsa) did not appear as authors on the paper.

Yes the author list is correct. Johan van der Vulgt is, in fact, an author on the paper. Juha and Richard are both in the acknowledgements, with Saleh. All present authors on this paper meet all authorship guidelines set by [ICMJE](https://www.icmje.org/recommendations/browse/roles-and-responsibilities/defining-the-role-of-authors-and-contributors.html). Notably, they all helped to prepare/write/edit the JOSS paper, and they all helped to prepare/write/edit documentation, and they all are signing on to take responsibility for the JOSS paper and supporting code/materials. These authors are also all following the review process and will help to make/verify project changes in accordance with the review critiques.

This being said - we have had a few very devoted developers join the project *after* the JOSS draft was initiated. It means there could be room to expand the acknowledgements list, if the JOSS editors allow such modifications. @Fei-Tao, can we extend the acknowledgements list to developers who joined *after* the JOSS draft was initiated?

> I noticed that the paper covers this aspect clearly but it is not so obvious on the website / github. Maybe it will help to expand on this in the website / github as well.

That is true - I am submitting a PR to add a statement of need to the general FreqAI introduction page so that documentation readers can see the objectives of the software. PR located here #7683

> Line 25, "the data is freely available", could the authors add some references / links to get access to such data?

Good point. We included the link to the open-sourced API that Freqtrade uses to freely collect, store, and update data for users (CCXT).  In direct response to your line here: all historical data is accessed via the `freqtrade download-data` command (which queries the linked API on the backend), where the users can indicate the exchange, timerange, timeframes, etc. Freqtrade then collects and stores that free data to disk which is then accessed by FreqAI for training/inferencing.

For real-time modeling, the free live data is downloaded/stored/updated automatically by FreqAI.

> Line 55 "novel methods which are not available anywhere else..." This sound like a very strong claim. Could the authors provide more details around the unique features (going deeper into the bullet points).

We would like to elaborate more on these methods - in an effort to avoid increasing the character count in the JOSS paper, we have embedded the permalinks to supporting material for each of the bullet points (Line 58-63). Here is the new section:

"
## Furnishing novel methods and features

Beyond the industry standard methods available through external libraries - `FreqAI` includes novel methods which are not available anywhere else in the open-source (or scientific) world. For example, `FreqAI` provides :

* a custom algorithm/methodology for adaptive modeling details [here](https://www.freqtrade.io/en/stable/freqai/#general-approach) and [here](https://www.freqtrade.io/en/stable/freqai-developers/#project-architecture)
* rapid and self-monitored feature engineering tools, details [here](https://www.freqtrade.io/en/stable/freqai-feature-engineering/#feature-engineering)
* unique model features/indicators, such as the [inlier metric](https://www.freqtrade.io/en/stable/freqai-feature-engineering/#inlier-metric)
* optimized data collection/storage algorithms, all code shown [here](https://github.com/freqtrade/freqtrade/blob/develop/freqtrade/freqai/data_drawer.py)
* safely integrated outlier detection methods, details [here](https://www.freqtrade.io/en/stable/freqai-feature-engineering/#outlier-detection)
* websocket communicated forecasts, details [here](https://www.freqtrade.io/en/stable/producer-consumer/)

Of particular interest for researchers, `FreqAI` provides the option of large scale experimentation via an optimized [websocket communications interface](https://www.freqtrade.io/en/stable/producer-consumer/).
"

> Line 182 "publication in scientific journal..." Could the authors add more details / links here?

We are still generating the results and preparing the submission. This means we do not yet have a link to a published article. We have change the sentence to say:

"Results from these studies are planned for submission to scientific journals as well as more general data science blogs (e.g. Medium)."

> Line 193 "FreqAI has no official sponsers", are there some unofficial sponsers? Wouldn't Github sponsorships count ?

That is correct, the only unofficial sponsors are the people who sponsor through github sponsors, as we already indicated in the paper. Should we change the wording? Please advise.
