### 2022-10-31 c3c6733
add statement of need to FreqAI docs
