### 2022-11-01 freqtrade/freqtrade#7668 @Fei-Tao
Hi @shagunsodhani, thank you for your detailed comments. Hi @robcaulk, it is okay to extend the acknowledgements list after the submission.
