### 2022-11-01 #4864 @Fei-Tao
Hi @shagunsodhani, thank you for your reviews. JOSS allows the co-publication of science, methods, and software. As long as the [FreqAI](https://github.com/freqtrade/freqtrade/tree/develop/freqtrade/freqai) itself reflects the substantial scholarly effort, we can accept it.
