### 2022-11-01 1975e94
Add test for no remaining
(kucoin case - https://github.com/freqtrade/freqtrade/issues/7757).
