### 2022-11-02 255eb71
start tracking the current candle in FreqAI, add robustness to corr_df caching and inference timer, add test for cache corr_df
