+++ freqtrade/freqai/data_drawer.py
@@ -98,6 +98,7 @@ class FreqaiDataDrawer:
                 "model_filename": "", "trained_timestamp": 0,
                 "data_path": "", "extras": {}}
         self.metric_tracker: Dict[str, Dict[str, Dict[str, list]]] = {}
+        self.current_candle: datetime = datetime.fromtimestamp(637887600)
 
     def update_metric_tracker(self, metric: str, value: float, pair: str) -> None:
         """
@@ -635,6 +636,8 @@ class FreqaiDataDrawer:
                         ignore_index=True,
                         axis=0,
                     )
+
+            self.current_candle = history_data[dk.pair][self.config['timeframe']].iloc[-1]['date']
 
     def load_all_pair_histories(self, timerange: TimeRange, dk: FreqaiDataKitchen) -> None:
         """
