### 2022-11-02 23b6915
fix issue with different backtesting prediction size
