### 2022-11-02 e25dea7
update channel disconnecting
