### 2022-11-02 c2bdaea
change exception handling in channel send
