### 2022-11-02 55bf195
remove debugging log calls
