### 2022-11-02 2dc55e8
better error handling channel send
