### 2022-11-02 cbede2e
refactor channel.send to avoid queue.put
