### 2022-11-03 000b0c2
prevent memory leaks from error in _broadcast_queue_data
