### 2022-11-03 b749f3e
add latency measure from ping in emc and ws_client
