### 2022-11-03 ff619ed
Improve explanation comment as to why we're waiting ourselfs
