### 2022-11-03 356d79b
verify mean and std exists in model metadata
