### 2022-11-03 6c4bdb8
remove special characters from feature names
