### 2022-11-03 6938ed6
change default conv_width to 1
