### 2022-11-03 90c5bfb
add default conv_width
