### 2022-11-04 0888b53
Udpate current_time handling for detail loop
