### 2022-11-04 e08cd6e
Reorder data cleaning
