### 2022-11-04 19d90b8
improve readibility in start_backtesting()
