### 2022-11-04 3ccc120
add option to force single precision
