### 2022-11-04 d4cfcbd
move write_metrics_to_disk to proper place in param table
