### 2022-11-04 257c833
add doc for single precision, dont allow half precision, add test
