+++ freqtrade/freqai/data_kitchen.py
@@ -1357,22 +1357,11 @@ class FreqaiDataKitchen:
 
         for col in df.columns[1:]:
             col_type = df[col].dtype
-
             if col_type != object:
-                c_min = df[col].min()
-                c_max = df[col].max()
                 if str(col_type)[:3] == "int":
-                    if c_min > np.iinfo(np.int8).min and c_max < np.iinfo(np.int8).max:
-                        df[col] = df[col].astype(np.int8)
-                    elif c_min > np.iinfo(np.int16).min and c_max < np.iinfo(np.int16).max:
-                        df[col] = df[col].astype(np.int16)
-                    elif c_min > np.iinfo(np.int32).min:
-                        df[col] = df[col].astype(np.int32)
+                    df[col] = df[col].astype(np.int32)
                 else:
-                    if c_min > np.finfo(np.float16).min and c_max < np.finfo(np.float16).max:
-                        df[col] = df[col].astype(np.float16)
-                    elif c_min > np.finfo(np.float32).min:
-                        df[col] = df[col].astype(np.float32)
+                    df[col] = df[col].astype(np.float32)
 
         end_mem = df.memory_usage().sum() / 1024**2
         print("Memory usage after optimization is: {:.2f} MB".format(end_mem))
