+++ user_data/strategies/wao/brain_util.py
@@ -2,7 +2,7 @@ import threading
 import watchdog
 import os
 import time
-from wao.brain_config import BrainConfig
+from wao.brain_config import *
 from wao._429_watcher import _429_Watcher
 from wao.error_watcher import Error_Watcher
 import pickle
@@ -14,15 +14,15 @@ from execution.backtest_signal import BacktestSignal
 
 def write_to_backtest_table(timestamp, coin, brain, time_out_hours, dup, type):
     print("STEP [1]++++++++++++++++++++++++++++++++++++" + ", write_to_backtest_table")
-    BrainConfig.BACKTEST_SIGNAL_LIST.append(BacktestSignal(brain, coin, type, time_out_hours, dup, timestamp=timestamp))
-    pickle.dump(BrainConfig.BACKTEST_SIGNAL_LIST, open(BrainConfig.BACKTEST_SIGNAL_LIST_PICKLE_FILE_PATH, 'wb'))
+    BACKTEST_SIGNAL_LIST.append(BacktestSignal(brain, coin, type, time_out_hours, dup, timestamp=timestamp))
+    pickle.dump(BACKTEST_SIGNAL_LIST, open(BACKTEST_SIGNAL_LIST_PICKLE_FILE_PATH, 'wb'))
 
 
 def perform_execute_buy(coin, brain, time_out_hours, dup):
     is_test_mode = False
-    if BrainConfig.MODE == Config.MODE_TEST:
+    if MODE == Config.MODE_TEST:
         is_test_mode = True
-    elif BrainConfig.MODE == Config.MODE_PROD:
+    elif MODE == Config.MODE_PROD:
         is_test_mode = False
 
     Config.COIN = coin
@@ -31,22 +31,22 @@ def perform_execute_buy(coin, brain, time_out_hours, dup):
     Config.ROMEO_D_UP_PERCENTAGE = dup
 
     romeo = Romeo.instance(is_test_mode, True)
-    BrainConfig.ROMEO_POOL[coin] = romeo
+    ROMEO_POOL[coin] = romeo
     romeo.start()
 
 
 def perform_execute_sell(coin):
     if Config.IS_SS_ENABLED:
-        romeo = BrainConfig.ROMEO_POOL.get(coin)
+        romeo = ROMEO_POOL.get(coin)
         if romeo is not None:
             romeo.perform_sell_signal(RomeoExitPriceType.SS)
 
 
 def perform_create_429_watcher():
-    print("perform_create_429_watcher: watching:- " + str(BrainConfig._429_DIRECTORY))
+    print("perform_create_429_watcher: watching:- " + str(_429_DIRECTORY))
     event_handler = _429_Watcher()
     observer = watchdog.observers.Observer()
-    observer.schedule(event_handler, path=BrainConfig._429_DIRECTORY, recursive=True)
+    observer.schedule(event_handler, path=_429_DIRECTORY, recursive=True)
     observer.start()
     try:
         while True:
@@ -57,10 +57,10 @@ def perform_create_429_watcher():
 
 
 def perform_create_error_watcher():
-    print("perform_create_error_watcher: watching:- " + str(BrainConfig._WAO_LOGS_DIRECTORY))
+    print("perform_create_error_watcher: watching:- " + str(_WAO_LOGS_DIRECTORY))
     event_handler = Error_Watcher()
     observer = watchdog.observers.Observer()
-    observer.schedule(event_handler, path=BrainConfig._WAO_LOGS_DIRECTORY, recursive=True)
[…]
