+++ user_data/strategies/bb_scalp_15m.py
@@ -6,17 +6,17 @@ import talib.abstract as ta
 import freqtrade.vendor.qtpylib.indicators as qtpylib
 import numpy  # noqa
 from wao.wao_strategy import WAOStrategy
-from wao.brain_config import BrainConfig
+from wao.brain_config import *
 
 
 class bb_scalp_15m(WAOStrategy):
     timeframe = '15m'
-    BrainConfig.BRAIN = "Freq_bb_15m"
+    BRAIN = "Freq_bb_15m"
 
     def __init__(self, config: dict):
         self.coin = str(config.get('pairs')[0]).split('/')[0]
         if self.coin == 'BTC' or self.coin == 'ADA':
-            BrainConfig.BRAIN = "Freq_bb_15m_ada_btc"
+            BRAIN = "Freq_bb_15m_ada_btc"
         super().__init__(config, 4, 0.14)
 
     # minimal_roi = {
