+++ user_data/strategies/wao/freq_data_cleaner.py
@@ -3,12 +3,12 @@ import datetime
 import time
 import os
 import sys
-from wao.brain_config import BrainConfig
+from wao.brain_config import *
 
 
 def clean_json():
-    print("clean_json: json_path = " + BrainConfig.BACKTEST_DOWNLOADED_JSON_DATA_FILE_PATH)
-    file = open(BrainConfig.BACKTEST_DOWNLOADED_JSON_DATA_FILE_PATH)
+    print("clean_json: json_path = " + BACKTEST_DOWNLOADED_JSON_DATA_FILE_PATH)
+    file = open(BACKTEST_DOWNLOADED_JSON_DATA_FILE_PATH)
     list = []
     data = json.load(file)
     for datas in data:
@@ -18,8 +18,8 @@ def clean_json():
         date = datetime.datetime.strptime(str(date), "%Y-%m-%d %H:%M:%S")
         year = date.year
         month = date.month
-        if year == int(BrainConfig.BACKTEST_DATA_CLEANER_YEAR) and month == int(
-                BrainConfig.BACKTEST_DATA_CLEANER_MONTH_INDEX) + 1:
+        if year == int(BACKTEST_DATA_CLEANER_YEAR) and month == int(
+                BACKTEST_DATA_CLEANER_MONTH_INDEX) + 1:
             list.append(datas)
     json_object = json.dumps(list)
     file.close()
@@ -27,10 +27,10 @@ def clean_json():
 
 
 def write_to_json(json_object):
-    print("write_to_json: json_path = " + BrainConfig.BACKTEST_DOWNLOADED_JSON_DATA_FILE_PATH)
+    print("write_to_json: json_path = " + BACKTEST_DOWNLOADED_JSON_DATA_FILE_PATH)
     with open("temp.json", "w") as outfile:
         outfile.write(json_object)
-    os.rename("temp.json", BrainConfig.BACKTEST_DOWNLOADED_JSON_DATA_FILE_PATH)
+    os.rename("temp.json", BACKTEST_DOWNLOADED_JSON_DATA_FILE_PATH)
 
 
 if len(sys.argv) < 2:
@@ -38,5 +38,5 @@ if len(sys.argv) < 2:
     python3 freq_data_cleaner.py [json_file] 
     """)
 else:
-    BrainConfig.BACKTEST_DOWNLOADED_JSON_DATA_FILE_PATH = sys.argv[1]
+    BACKTEST_DOWNLOADED_JSON_DATA_FILE_PATH = sys.argv[1]
     clean_json()
