+++ user_data/strategies/Hacklemore.py
@@ -10,7 +10,7 @@ from freqtrade.persistence import Trade
 from technical.indicators import RMI
 from statistics import mean
 from wao.wao_strategy import WAOStrategy
-from wao.brain_config import BrainConfig
+from wao.brain_config import *
 
 """
 TODO: 
@@ -20,7 +20,7 @@ TODO:
 
 
 class Hacklemore(WAOStrategy):
-    BrainConfig.BRAIN = "Freq_Hacklemore"
+    BRAIN = "Freq_Hacklemore"
 
     def __init__(self, config: dict):
         super().__init__(config, 8, 0.15)
