+++ user_data/strategies/bb_scalp_15m.py
@@ -6,7 +6,7 @@ import talib.abstract as ta
 import freqtrade.vendor.qtpylib.indicators as qtpylib
 import numpy  # noqa
 from wao.wao_strategy import WAOStrategy
-from wao.brain_config import *
+from wao.brain_config2 import *
 
 
 class bb_scalp_15m(WAOStrategy):
