+++ user_data/strategies/Strategy003.py
@@ -8,7 +8,7 @@ import freqtrade.vendor.qtpylib.indicators as qtpylib
 import numpy  # noqa
 
 from wao.wao_strategy import WAOStrategy
-from wao.brain_config2 import *
+from wao.brain_config import *
 
 
 class Strategy003(WAOStrategy):
