+++ user_data/strategies/Cluc5werk.py
@@ -5,7 +5,7 @@ from freqtrade.strategy.interface import IStrategy
 from freqtrade.strategy import merge_informative_pair
 from pandas import DataFrame
 from wao.wao_strategy import WAOStrategy
-from wao.brain_config2 import *
+from wao.brain_config import *
 
 
 def bollinger_bands(stock_price, window_size, num_of_std):
