+++ user_data/strategies/Strategy002.py
@@ -7,11 +7,11 @@ import freqtrade.vendor.qtpylib.indicators as qtpylib
 import numpy # noqa
 
 from wao.wao_strategy import WAOStrategy
-from wao.brain_config import *
+from wao.brain_config import BrainConfig
 
 
 class Strategy002(WAOStrategy):
-    BRAIN = 'Freq_Strategy002'
+    BrainConfig.BRAIN = 'Freq_Strategy002'
 
     def __init__(self, config: dict):
         super().__init__(config, 10, 0.001)
