+++ user_data/strategies/wao/wao_strategy_controller2.py
@@ -7,7 +7,7 @@ from notifier import send_start_deliminator_message
 from execution.config import Config
 
 
-class WAOStrategyController:
+class WAOStrategyController2:
 
     def __init__(self, time_out_hours, dup):
         self.time_out_hours = time_out_hours
