+++ user_data/strategies/wao/wao_strategy_futures.py
@@ -2,7 +2,7 @@ from datetime import datetime
 from typing import Optional
 from freqtrade.persistence import Trade
 from freqtrade.strategy import IStrategy
-from wao_strategy_controller import WAOStrategyController
+from wao_strategy_controller2 import WAOStrategyController2
 
 
 class WAOStrategy_futures(IStrategy):
@@ -16,7 +16,7 @@ class WAOStrategy_futures(IStrategy):
 
     def __init__(self, config: dict, time_out_hours, dup):
         super().__init__(config)
-        self.controller = WAOStrategyController(time_out_hours, dup)
+        self.controller = WAOStrategyController2(time_out_hours, dup)
 
     def confirm_trade_entry(self, pair: str, order_type: str, amount: float, rate: float,
                             time_in_force: str, current_time: datetime, entry_tag: Optional[str],
