+++ user_data/strategies/wao/stop_bot.py
@@ -5,7 +5,7 @@ from execution.config import Config
 from execution.broker.binance_future_broker import Binance_Future_Broker
 from execution.broker.binance_spot_broker import Binance_Spot_Broker
 from execution.notifier import Notifier
-from notifier import send_stop_bot_message
+from wao.notifier import send_stop_bot_message
 
 
 if len(sys.argv) < 3:
