+++ user_data/strategies/wao/notifier.py
@@ -1,6 +1,4 @@
-import sys
-sys.path.append("wao/")
-from brain_config import BrainConfig
+from wao.brain_config import BrainConfig
 import requests
 
 from execution.config import Config
