### 2022-11-05 53df607
avoid duplicating features with okx/gateio, ensure inference timer gets logged
