+++ freqtrade/freqai/data_kitchen.py
@@ -1177,7 +1177,7 @@ class FreqaiDataKitchen:
                     ready for training
         """
         pairs = self.freqai_config["feature_parameters"].get("include_corr_pairlist", [])
-
+        current_pair = current_pair.replace(':', '')
         for pair in pairs:
             pair = pair.replace(':', '')  # lightgbm doesnt work with colons
             if current_pair != pair:
