### 2022-11-05 43bdd34
Optimize reduce_dataframe_footprint function
