### 2022-11-05 5bd3e54
Add test for detail backtesting
