### 2022-11-05 d089fda
Fix current-time_det calculation
