### 2022-11-05 ded57fb
Remove no longer valid test part
