### 2022-11-06 f9524ae
Improve temporary log output for exchange_ws
