+++ requirements.txt
@@ -12,7 +12,7 @@ arrow==1.2.3
 cachetools==4.2.2
 requests==2.28.1
 urllib3==1.26.12
-jsonschema==4.16.0
+jsonschema==4.17.0
 TA-Lib==0.4.25
 technical==1.3.0
 tabulate==0.9.0
