+++ docs/requirements-docs.txt
@@ -1,5 +1,5 @@
 markdown==3.3.7
-mkdocs==1.4.1
+mkdocs==1.4.2
 mkdocs-material==8.5.7
 mdx_truly_sane_lists==1.3
 pymdown-extensions==9.7
