+++ requirements.txt
@@ -46,7 +46,7 @@ psutil==5.9.3
 colorama==0.4.6
 # Building config files interactively
 questionary==1.10.0
-prompt-toolkit==3.0.31
+prompt-toolkit==3.0.32
 # Extensions to datetime library
 python-dateutil==2.8.2
 
