### 2022-11-07 d978ff6
Bump orjson from 3.8.0 to 3.8.1
Bumps [orjson](https://github.com/ijl/orjson) from 3.8.0 to 3.8.1.
- [Release notes](https://github.com/ijl/orjson/releases)
- [Changelog](https://github.com/ijl/orjson/blob/master/CHANGELOG.md)
- [Commits](https://github.com/ijl/orjson/compare/3.8.0...3.8.1)

---
updated-dependencies:
- dependency-name: orjson
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
