+++ requirements.txt
@@ -29,7 +29,7 @@ py_find_1st==1.1.5
 # Load ticker files 30% faster
 python-rapidjson==1.9
 # Properly format api responses
-orjson==3.8.0
+orjson==3.8.1
 
 # Notify systemd
 sdnotify==0.3.2
