### 2022-11-07 60cfda5
Add very basic exception handling
