+++ user_data/strategies/wao/stop_bot.py
@@ -6,15 +6,16 @@ from execution.broker.binance_future_broker import Binance_Future_Broker
 from execution.broker.binance_spot_broker import Binance_Spot_Broker
 from execution.notifier import Notifier
 
-if len(sys.argv) < 3:
+if len(sys.argv) != 5:
     exit("""Incorrect number of arguments. 
     python3 stop_bot.py [mode:test/prod] [brain] [coin] [reason]
     """)
 else:
     is_test_mode = sys.argv[1] == "test" or sys.argv[1] == "True"
-    Config.BRAIN = str(sys.argv[2]) if str(sys.argv[2]).split("_")[0] == "Freq" else "Freq_" + str(sys.argv[2])
-    coin = sys.argv[2]
-    reason = str(sys.argv[4]).replace("_", "#") + " " + coin
+    brain = str(sys.argv[2])
+    Config.BRAIN = brain if brain.split("_")[0] == "Freq" else "Freq_" + brain
+    Config.COIN = sys.argv[3]
+    reason = str(sys.argv[4]).replace("_", "#") + " " + Config.COIN
 
     notifier = Notifier(is_test_mode)
 
@@ -23,7 +24,7 @@ else:
     else:
         broker = Binance_Spot_Broker(notifier)
 
-    broker.close_all_open_positions_by_coin(coin)
+    broker.close_all_open_positions_by_coin()
 
     notifier.send_stop_bot_message(reason)
 
