### 2022-11-08 8d9988a
enforce date column in backtesting freqai predictions files
