### 2022-11-08 3e57c18
add fix_live_predictions function to backtesting
