### 2022-11-09 8ee95db
refactoring backtesting_fit_live_predictions function
