### 2022-11-09 d84ece7
Use conditional orders for stop orders
