### 2022-11-10 9e17eab
Improve Bybit configuration
