### 2022-11-10 7147f52
FreqAI also requires plotting dependencies
cloess #7726
