### 2022-11-10 3e676db
Add properties to simplify timerange handling
