### 2022-11-10 57313dd
Update some usages of timerange to new, simplified method
