+++ freqtrade/freqai/utils.py
@@ -230,7 +230,5 @@ def get_timerange_backtest_live_models(config: Config) -> str:
     dk = FreqaiDataKitchen(config)
     models_path = dk.get_full_models_path(config)
     timerange, _ = dk.get_timerange_and_assets_end_dates_from_ready_models(models_path)
-    start_date = datetime.fromtimestamp(timerange.startts, tz=timezone.utc)
-    end_date = datetime.fromtimestamp(timerange.stopts, tz=timezone.utc)
-    tr = f"{start_date.strftime('%Y%m%d')}-{end_date.strftime('%Y%m%d')}"
+    tr = f"{timerange.startdt.strftime('%Y%m%d')}-{timerange.stopdt.strftime('%Y%m%d')}"
     return tr
