### 2022-11-10 0f9c5f8
Simplify timerange handling
