### 2022-11-11 0541339
fix loading of metric tracker from disk
