### 2022-11-11 e46a57b
move mem logs to debug
