### 2022-11-11 aef0324
set markets for ws exchange on reload
