### 2022-11-12 6746868
store dataprovider to self instead of strategy
