### 2022-11-12 9c6b97c
ensure normalization acceleration methods are employed in RL
