### 2022-11-12 8c092d4
Update README.md
Removed the FTX option under the "Supported Exchange marketplaces" title, because of FTX bankrupt.
