+++ freqtrade/freqai/utils.py
@@ -229,7 +229,12 @@ def get_timerange_backtest_live_models(config: Config) -> str:
     """
     dk = FreqaiDataKitchen(config)
     models_path = dk.get_full_models_path(config)
-    timerange, _ = dk.get_timerange_and_assets_end_dates_from_ready_models(models_path)
+    timerange: TimeRange = TimeRange()
+    if not config.get("save_live_data_backtest", True):
+        timerange, _ = dk.get_timerange_and_assets_end_dates_from_ready_models(models_path)
+    else:
+        timerange = dk.get_timerange_from_backtesting_live_dataframe()
+
     start_date = datetime.fromtimestamp(timerange.startts, tz=timezone.utc)
     end_date = datetime.fromtimestamp(timerange.stopts, tz=timezone.utc)
     tr = f"{start_date.strftime('%Y%m%d')}-{end_date.strftime('%Y%m%d')}"
