### 2022-11-12 e71a8b8
add ability to integrate state info or not, and prevent state info integration during backtesting
