### 2022-11-12 954da4f
Add "forcebuy error" exception log
part of #7727
