+++ freqtrade/rpc/telegram.py
@@ -1061,6 +1061,7 @@ class Telegram(RPCHandler):
             try:
                 self._rpc._rpc_force_entry(pair, price, order_side=order_side)
             except RPCException as e:
+                logger.exception("Forcebuy error!")
                 self._send_msg(str(e))
 
     def _force_enter_inline(self, update: Update, _: CallbackContext) -> None:
