### 2022-11-12 914bdbd
Remove FTX from list of supported exchanges
