### 2022-11-13 535c365
Disable ftx ccxt_compat tests
Their API is down due to being insolvent ...
