### 2022-11-13 81f800a
switch to using FT calc_profi_pct, reverse entry/exit fees
