### 2022-11-13 942840d
Improve setting wording to keep future possibilities open
