### 2022-11-13 af9e400
add test coverage, fix bug in base environment. Ensure proper fee is used.
