### 2022-11-13 1e9e788
fix constants.py, fix freqai test
