### 2022-11-13 3c249ba
add doc for data_kitchen_thread_count`
