### 2022-11-13 388ca21
update docs, fix bug in environment
