+++ docs/freqai-reinforcement-learning.md
@@ -16,7 +16,10 @@ Reinforcement learning is a natural progression for FreqAI, since it adds a new 
 
 ### The RL interface
 
-With the current framework, we aim to expose the training environment to the user via the common "prediction model" file (i.e. CatboostClassifier, LightGBMRegressor, etc.). Users inherit our base environment in this file, which allows them to override as much or as little of the environment as they wish. 
+With the current framework, we aim to expose the training environment via the common "prediction model" file, which is a user inherited `BaseReinforcementLearner` object (e.g. `freqai/prediction_models/ReinforcementLearner`). Inside this user class, the RL environment is available and customized via `MyRLEnv`:
+
+
+
 
 We envision the majority of users focusing their effort on creative design of the `calculate_reward()` function [details here](#creating-the-reward), while leaving the rest of the environment untouched. Other users may not touch the environment at all, and they will only play with the configruation settings and the powerful feature engineering that already exists in FreqAI. Meanwhile, we enable advanced users to create their own model classes entirely.
 
@@ -49,7 +52,7 @@ where `ReinforcementLearner` will use the templated `ReinforcementLearner` from 
             informative[f"%-{pair}mfi-period_{t}"] = ta.MFI(informative, timeperiod=t)
             informative[f"%-{pair}adx-period_{t}"] = ta.ADX(informative, window=t)
 
-        # The following features are necessary for RL models
+        # The following raw price values are necessary for RL models
         informative[f"%-{pair}raw_close"] = informative["close"]
         informative[f"%-{pair}raw_open"] = informative["open"]
         informative[f"%-{pair}raw_high"] = informative["high"]
@@ -131,11 +134,12 @@ It is important to consider that `&-action` depends on which environment they ch
 
 ## Configuring the Reinforcement Learner
 
-In order to configure the `Reinforcement Learner` the following dictionary to their `freqai` config:
+In order to configure the `Reinforcement Learner` the following dictionary must exist in the `freqai` config:
 
 ```json
         "rl_config": {
             "train_cycles": 25,
+            "add_state_info": true,
             "max_trade_duration_candles": 300,
             "max_training_drawdown_pct": 0.02,
             "cpu_count": 8,
@@ -148,11 +152,14 @@ In order to configure the `Reinforcement Learner` the following dictionary to th
         }
 ```
 
-Parameter details can be found [here](freqai-parameter-table.md), but in general the `train_cycles` decides how many times the agent should cycle through the candle data in its artificial environemtn to train weights in the model. `model_type` is a string which selects one of the available models in [stable_baselines](https://stable-baselines3.readthedocs.io/en/master/)(external link). 
+Parameter details can be found [here](freqai-parameter-table.md), but in general the `train_cycles` decides how many times the agent should cycle through the candle data in its artificial environment to train weights in the model. `model_type` is a string which selects one of the available models in [stable_baselines](https://stable-baselines3.readthedocs.io/en/master/)(external link). 
+
+!!! Note
+    Remember that the general `model_training_parameters` dictionary should contain all the model hyperparameter customizations for the particular `model_type`. For example, `PPO` parameters can be found [here](https://stable-baselines3.readthedocs.io/en/master/modules/ppo.html).
 
 ## Creating the reward
 
[…]
