### 2022-11-13 b421521
help default ReinforcementLearner users by assigning the model_type automatically
