### 2022-11-13 c76afc2
explain how to choose environments, and how to customize them
