### 2022-11-13 c8d3e57
add note that these environments are designed for short-long bots only.
