### 2022-11-13 bf4d5b4
ensure model_type is defined
