+++ docs/requirements-docs.txt
@@ -2,5 +2,5 @@ markdown==3.3.7
 mkdocs==1.4.2
 mkdocs-material==8.5.8
 mdx_truly_sane_lists==1.3
-pymdown-extensions==9.7
+pymdown-extensions==9.8
 jinja2==3.1.2
