+++ requirements.txt
@@ -40,7 +40,7 @@ pydantic==1.10.2
 uvicorn==0.19.0
 pyjwt==2.6.0
 aiofiles==22.1.0
-psutil==5.9.3
+psutil==5.9.4
 
 # Support for colorized terminal output
 colorama==0.4.6
