### 2022-11-14 7275d48
Bump ccxt from 2.1.54 to 2.1.75
Bumps [ccxt](https://github.com/ccxt/ccxt) from 2.1.54 to 2.1.75.
- [Release notes](https://github.com/ccxt/ccxt/releases)
- [Changelog](https://github.com/ccxt/ccxt/blob/master/exchanges.cfg)
- [Commits](https://github.com/ccxt/ccxt/compare/2.1.54...2.1.75)

---
updated-dependencies:
- dependency-name: ccxt
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
