### 2022-11-14 a951b49
Use Generator when sending initial dataframes
