+++ .pre-commit-config.yaml
@@ -17,7 +17,7 @@ repos:
           - types-filelock==3.2.7
           - types-requests==2.28.11.4
           - types-tabulate==0.9.0.0
-          - types-python-dateutil==2.8.19.2
+          - types-python-dateutil==2.8.19.3
         # stages: [push]
 
   - repo: https://github.com/pycqa/isort
