### 2022-11-14 30b4679
Delist FTX, following ccxt's delisting.
