### 2022-11-14 926f870
Improve the training by randomly setting the starting point
This is an improvement for all kind of RL stuff, the starting point sets in the first third of the data which ensures that the model not only optimizes the policy or Q values to the choosen timerange.
