+++ freqtrade/rpc/api_server/webserver.py
@@ -1,7 +1,6 @@
 import asyncio
 import logging
 from ipaddress import IPv4Address
-from threading import Thread
 from typing import Any, Dict
 
 import orjson
@@ -15,7 +14,7 @@ from starlette.responses import JSONResponse
 from freqtrade.constants import Config
 from freqtrade.exceptions import OperationalException
 from freqtrade.rpc.api_server.uvicorn_threaded import UvicornServer
-from freqtrade.rpc.api_server.ws import ChannelManager
+from freqtrade.rpc.api_server.ws.message_stream import MessageStream
 from freqtrade.rpc.api_server.ws_schemas import WSMessageSchemaType
 from freqtrade.rpc.rpc import RPC, RPCException, RPCHandler
 
@@ -51,9 +50,10 @@ class ApiServer(RPCHandler):
     # Exchange - only available in webserver mode.
     _exchange = None
     # websocket message queue stuff
-    _ws_channel_manager = None
-    _ws_thread = None
-    _ws_loop = None
+    # _ws_channel_manager = None
+    # _ws_thread = None
+    # _ws_loop = None
+    _message_stream = None
 
     def __new__(cls, *args, **kwargs):
         """
@@ -71,14 +71,15 @@ class ApiServer(RPCHandler):
             return
         self._standalone: bool = standalone
         self._server = None
+
         self._ws_queue = None
-        self._ws_background_task = None
+        self._ws_publisher_task = None
 
         ApiServer.__initialized = True
 
         api_config = self._config['api_server']
 
-        ApiServer._ws_channel_manager = ChannelManager()
+        # ApiServer._ws_channel_manager = ChannelManager()
 
         self.app = FastAPI(title="Freqtrade API",
                            docs_url='/docs' if api_config.get('enable_openapi', False) else None,
@@ -107,18 +108,18 @@ class ApiServer(RPCHandler):
             logger.info("Stopping API Server")
             self._server.cleanup()
 
-        if self._ws_thread and self._ws_loop:
-            logger.info("Stopping API Server background tasks")
+        # if self._ws_thread and self._ws_loop:
+        #     logger.info("Stopping API Server background tasks")
 
-            if self._ws_background_task:
-                # Cancel the queue task
-                self._ws_background_task.cancel()
+        #     if self._ws_background_task:
+        #         # Cancel the queue task
+        #         self._ws_background_task.cancel()
 
-            self._ws_thread.join()
+        #     self._ws_thread.join()
 
-        self._ws_thread = None
-        self._ws_loop = None
-        self._ws_background_task = None
+        # self._ws_thread = None
+        # self._ws_loop = None
+        # self._ws_background_task = None
 
     @classmethod
     def shutdown(cls):
@@ -170,50 +171,101 @@ class ApiServer(RPCHandler):
         )
 
         app.add_exception_handler(RPCException, self.handle_rpc_exception)
+        app.add_event_handler(
+            event_type="startup",
+            func=self._api_startup_event
+        )
+        app.add_event_handler(
+            event_type="shutdown",
+            func=self._api_shutdown_event
+        )
 
-    def start_message_queue(self):
-        if self._ws_thread:
-            return
+    async def _api_startup_event(self):
+        if not ApiServer._message_stream:
+            ApiServer._message_stream = MessageStream()
 
-        # Create a new loop, as it'll be just for the background thread
-        self._ws_loop = asyncio.new_event_loop()
+        if not self._ws_queue:
+            self._ws_queue = ThreadedQueue()
 
-        # Start the thread
-        self._ws_thread = Thread(target=self._ws_loop.run_forever)
-        self._ws_thread.start()
+        if not self._ws_publisher_task:
[…]
