+++ freqtrade/rpc/api_server/ws/channel.py
@@ -125,223 +125,3 @@ class WebSocketChannel:
             yield self
         finally:
             await self.close()
-
-
-# class WebSocketChannel:
-#     """
-#     Object to help facilitate managing a websocket connection
-#     """
-
-#     def __init__(
-#         self,
-#         websocket: WebSocketType,
-#         channel_id: Optional[str] = None,
-#         drain_timeout: int = 3,
-#         throttle: float = 0.01,
-#         serializer_cls: Type[WebSocketSerializer] = HybridJSONWebSocketSerializer
-#     ):
-
-#         self.channel_id = channel_id if channel_id else uuid4().hex[:8]
-
-#         # The WebSocket object
-#         self._websocket = WebSocketProxy(websocket)
-
-#         self.drain_timeout = drain_timeout
-#         self.throttle = throttle
-
-#         self._subscriptions: List[str] = []
-#         # 32 is the size of the receiving queue in websockets package
-#         self.queue: asyncio.Queue[Dict[str, Any]] = asyncio.Queue(maxsize=32)
-#         self._relay_task = asyncio.create_task(self.relay())
-
-#         # Internal event to signify a closed websocket
-#         self._closed = asyncio.Event()
-
-#         # Wrap the WebSocket in the Serializing class
-#         self._wrapped_ws = serializer_cls(self._websocket)
-
-#     def __repr__(self):
-#         return f"WebSocketChannel({self.channel_id}, {self.remote_addr})"
-
-#     @property
-#     def raw_websocket(self):
-#         return self._websocket.raw_websocket
-
-#     @property
-#     def remote_addr(self):
-#         return self._websocket.remote_addr
-
-#     async def _send(self, data):
-#         """
-#         Send data on the wrapped websocket
-#         """
-#         await self._wrapped_ws.send(data)
-
-#     async def send(self, data) -> bool:
-#         """
-#         Add the data to the queue to be sent.
-#         :returns: True if data added to queue, False otherwise
-#         """
-
-#         # This block only runs if the queue is full, it will wait
-#         # until self.drain_timeout for the relay to drain the outgoing queue
-#         # We can't use asyncio.wait_for here because the queue may have been created with a
-#         # different eventloop
-#         start = time.time()
-#         while self.queue.full():
-#             await asyncio.sleep(1)
-#             if (time.time() - start) > self.drain_timeout:
-#                 return False
-
-#         # If for some reason the queue is still full, just return False
-#         try:
-#             self.queue.put_nowait(data)
-#         except asyncio.QueueFull:
-#             return False
-
-#         # If we got here everything is ok
-#         return True
-
-#     async def recv(self):
-#         """
-#         Receive data on the wrapped websocket
-#         """
-#         return await self._wrapped_ws.recv()
-
-#     async def ping(self):
-#         """
-#         Ping the websocket
-#         """
-#         return await self._websocket.ping()
-
-#     async def close(self):
-#         """
-#         Close the WebSocketChannel
-#         """
-
-#         try:
-#             await self.raw_websocket.close()
-#         except Exception:
-#             pass
-
-#         self._closed.set()
-#         self._relay_task.cancel()
-
-#     def is_closed(self) -> bool:
-#         """
-#         Closed flag
-#         """
-#         return self._closed.is_set()
-
-#     def set_subscriptions(self, subscriptions: List[str] = []) -> None:
-#         """
-#         Set which subscriptions this channel is subscribed to
-
-#         :param subscriptions: List of subscriptions, List[str]
-#         """
-#         self._subscriptions = subscriptions
-
-#     def subscribed_to(self, message_type: str) -> bool:
-#         """
-#         Check if this channel is subscribed to the message_type
-
[…]
