### 2022-11-15 ad56c45
Add total profit to the monitor file  and add monitor for the trainingf
This can help for beginners to see what is happening inside of the training, and helps them understand how much profit of their agent did under the period of training and eval.
