+++ user_data/strategies/bbrsi_scalp_futures_short.py
@@ -111,14 +111,6 @@ class bbrsi_scalp_futures_short(WAOStrategy_futures):
         return dataframe
 
     def populate_entry_trend(self, dataframe: DataFrame, metadata: dict) -> DataFrame:
-        dataframe.loc[
-            (
-                    (qtpylib.crossed_above(dataframe['rsi'], 30)) &  # Signal: RSI crosses above 30
-                    (dataframe['tema'] <= dataframe['bb_middleband']) &  # Guard
-                    (dataframe['tema'] > dataframe['tema'].shift(1)) &  # Guard
-                    (dataframe['volume'] > 0)  # Make sure Volume is not 0
-            ),
-            ['enter_long', 'enter_tag']] = (1, 'rsi_cross')
 
         dataframe.loc[
             (
@@ -132,14 +124,7 @@ class bbrsi_scalp_futures_short(WAOStrategy_futures):
         return dataframe
 
     def populate_exit_trend(self, dataframe: DataFrame, metadata: dict) -> DataFrame:
-        dataframe.loc[
-            (
-                    (qtpylib.crossed_above(dataframe['rsi'], 70)) &  # Signal: RSI crosses above 70
-                    (dataframe['tema'] > dataframe['bb_middleband']) &  # Guard
-                    (dataframe['tema'] < dataframe['tema'].shift(1)) &  # Guard
-                    (dataframe['volume'] > 0)  # Make sure Volume is not 0
-            ),
-            ['exit_long', 'exit_tag']] = (1, 'rsi_too_high')
+
         dataframe.loc[
             (
                     (qtpylib.crossed_below(dataframe['rsi'], 30)) &  # Signal: RSI crosses below 30
