+++ user_data/strategies/bbrsi_optimal_rsi_hyperopt.py
@@ -1,150 +1,0 @@
-# --- Do not remove these libs ---
-from pandas import DataFrame
-# --------------------------------
-
-import talib.abstract as ta
-import freqtrade.vendor.qtpylib.indicators as qtpylib
-import numpy  # noqa
-from wao.wao_strategy import WAOStrategy
-from freqtrade.strategy import (IntParameter)
-from functools import reduce
-
-
-class bbrsi_optimal_rsi_hyperopt(WAOStrategy):
-    brain = "Freq_bbrsi_scalp"
-
-    def __init__(self, config: dict):
-        self.coin = str(config.get('pairs')[0]).split('/')[0]
-        if self.coin == 'BTC' or self.coin == 'ADA':
-            self.brain = "Freq_bbrsi_scalp_ada_btc"
-        super().__init__(config, self.brain, 8, 0.15)
-    # buy hyperspace params
-    buy_params = {
-        "buy_rsi": 50,
-    }
-
-    # Sell hyperspace params:
-    sell_params = {
-        "sell_rsi": 60,
-    }
-
-    # Minimal ROI designed for the strategy.
-    # This attribute will be overridden if the config file contains "minimal_roi"
-    minimal_roi = {
-        "190": 0.005,  # Exit after 500 minutes there is at least 0.5% profit
-        "175": 0.006,  # Exit after 500 minutes there is at least 0.5% profit
-        "160": 0.007,  # Exit after 40 minutes if there is at least 1% profit
-        "145": 0.008,  # Exit after 40 minutes if there is at least 1% profit
-        "130": 0.009,  # Exit after 40 minutes if there is at least 1% profit
-        "115": 0.010,  # Exit after 20 minutes if there is at least 1.5% profit
-        "100": 0.011,  # Exit after 20 minutes if there is at least 1.5% profit
-        "85": 0.012,  # Exit after 20 minutes if there is at least 1.5% profit
-        "60": 0.013,  # Exit after 20 minutes if there is at least 1.5% profit
-        "45": 0.014,  # Exit immediately if there is at least 2% profit
-        "30": 0.016,  # Exit immediately if there is at least 2% profit
-        "15": 0.018,  # Exit immediately if there is at least 2% profit
-        "0": 0.020,  # Exit immediately if there is at least 2% profit
-    }
-
-    # Optimal stoploss designed for the strategy
-    # This attribute will be overridden if the config file contains "stoploss"
-    stoploss = -0.01
-
-    # Optimal timeframe for the strategy
-    timeframe = '5m'
-
-    # trailing stoploss
-    trailing_stop = False
-    trailing_stop_positive = 0.01
-    trailing_stop_positive_offset = 0.02
-
-    # run "populate_indicators" only for new candle
-    process_only_new_candles = False
-
-    # Experimental settings (configuration will overide these if set)
-    use_sell_signal = True
-    sell_profit_only = True
-    ignore_roi_if_buy_signal = False
-
-    buy_rsi = IntParameter(0, 60, default=50, space="buy")
-    sell_rsi = IntParameter(50, 100, default=60, space="sell")
-
-    def informative_pairs(self):
-        """
-        Define additional, informative pair/interval combinations to be cached from the exchange.
-        These pair/interval combinations are non-tradeable, unless they are part
-        of the whitelist as well.
-        For more information, please consult the documentation
-        :return: List of tuples in the format (pair, interval)
-            Sample: return [("ETH/USDT", "5m"),
-                            ("BTC/USDT", "15m"),
-                            ]
-        """
-        return []
-
[…]
