+++ user_data/hyperopts/WAOHyperOptLoss.py
@@ -37,10 +37,7 @@ class WAOHyperOptLoss(IHyperOptLoss):
 
             x += 1
 
-        if trade_count == 0 or win_trades_number == 0:
-            win_trades_percentage = 0.0000001
-        else:
-            win_trades_percentage = win_trades_number / trade_count
+        win_trades_percentage = win_trades_number / trade_count
         """We can put that into a function above"""
         y = 0
         profit_sell_signal = []

[…]
