+++ user_data/strategies/bbrsi_scalp.py
@@ -79,6 +79,7 @@ class bbrsi_scalp(WAOStrategy):
         """
         Adds several different TA indicators to the given DataFrame
         Performance Note: For the best performance be frugal on the number of indicators
+
         you are using. Let uncomment only the indicator you are using in your refined-strategies
         or your hyperopt configuration, otherwise you will waste your memory and CPU usage.
         """
