### 2022-11-17 48e5a45
rpc_test: dont replicate whole response,
updating  what's changed improves readability
