### 2022-11-17 913749c
backtesting_from_live_saved_files - code refactoring
