### 2022-11-17 99bff9c
backtesting_from_live_saved_files - code refactoring
