### 2022-11-17 436b314
add safe_remaining
fixes #7757
