### 2022-11-17 875e9ab
change df serialization to avoid mem leak
