### 2022-11-17 60fcd8d
fix skipped mac test, fix RL bug in add_state_info, fix use of __import__, revise doc
