+++ user_data/strategies/wao/brain_util.py
@@ -12,13 +12,14 @@ from execution.romeo import Romeo, RomeoExitPriceType
 from execution.backtest_signal import BacktestSignal
 
 
-def write_to_backtest_table(timestamp, coin, brain, time_out_hours, dup, type):
+def write_to_backtest_table(timestamp, coin, dup, type):
     print("STEP [1]++++++++++++++++++++++++++++++++++++" + ", write_to_backtest_table")
-    BrainConfig.BACKTEST_SIGNAL_LIST.append(BacktestSignal(brain, coin, type, time_out_hours, dup, timestamp=timestamp))
+    BrainConfig.BACKTEST_SIGNAL_LIST.append(BacktestSignal(BrainConfig.BRAIN, coin, type,
+                                                           Config.ROMEO_SS_TIMEOUT_HOURS, dup, timestamp=timestamp))
     pickle.dump(BrainConfig.BACKTEST_SIGNAL_LIST, open(BrainConfig.BACKTEST_SIGNAL_LIST_PICKLE_FILE_PATH, 'wb'))
 
 
-def perform_execute_buy(coin, brain, time_out_hours, dup):
+def perform_execute_buy(coin, dup):
     is_test_mode = False
     if BrainConfig.MODE == Config.MODE_TEST:
         is_test_mode = True
@@ -26,8 +27,6 @@ def perform_execute_buy(coin, brain, time_out_hours, dup):
         is_test_mode = False
 
     Config.COIN = coin
-    Config.BRAIN = brain
-    Config.ROMEO_SS_TIMEOUT_HOURS = time_out_hours
     Config.ROMEO_D_UP_PERCENTAGE = dup
 
     romeo = Romeo.instance(is_test_mode, True)
