+++ requirements-freqai-rl.txt
@@ -6,4 +6,3 @@ torch==1.12.1
 stable-baselines3==1.6.1
 gym==0.21
 sb3-contrib==1.6.1
-tensorboard==2.10.1
