### 2022-11-18 b6a8e42
remove redundant timestamp conversion in ws serializer
