### 2022-11-18 0cb6f71
better error handling, true async sending, more readable api
