### 2022-11-19 d02da27
document the simplifications of the training environment
