### 2022-11-19 80d070e
update code to use historic_predictions for freqai_backtest_live_models
