### 2022-11-19 c1a73a5
move sleep call in send, minor cleanup
