### 2022-11-20 fdc82af
fix tests - update code to backtest with historic_predictions
