### 2022-11-20 48a1f24
update typing, remove unneeded try block, readd sleep
