### 2022-11-20 d2870d4
change typing to async iterator
