+++ docs/rest-api.md
@@ -458,3 +458,172 @@ The correct configuration for this case is `http://localhost:8080` - the main pa
 
 !!! Note
     We strongly recommend to also set `jwt_secret_key` to something random and known only to yourself to avoid unauthorized access to your bot.
+
+
+### Using SSL/TLS
+
+SSL/TLS is used to provide security and encrypt network traffic. Freqtrade does not directly support SSL, but you can easily accomplish this with a reverse proxy such as Nginx or Traefik. Below are some steps to help you get started on setting one up for your bot. For the sake of simplicity, we will use a native installation of Nginx and certbot.
+
+
+**Prerequisites**
+
+Before starting this tutorial, you will need a few things.
+
+- A Freqtrade bot set up and running
+- A registered domain name, e.g. myftbot.com
+- A DNS A record for the top level domain pointing to your server's public IP
+
+**Step 1: Installing Nginx and Certbot**
+
+Once you have all of the prerequisites, the first step is to get Nginx installed on your system. This tutorial assumes the use of Ubuntu 20.04, though you can find your linux distro's package commands via a search engine. First, update your local package index so that you have access to the most recent package listings then install Nginx:
+
+``` bash
+> sudo apt update
+> sudo apt install nginx
+```
+
+After accepting the installation, Nginx and any dependencies will be installed to your system and automatically started. You can check it is running with `systemd`:
+
+``` bash
+> sudo systemctl status nginx
+
+● nginx.service - A high performance web server and a reverse proxy server
+     Loaded: loaded (/lib/systemd/system/nginx.service; enabled; vendor preset: enabled)
+     Active: active (running) since Wed 2022-11-16 12:09:27 MST; 4 days ago
+       Docs: man:nginx(8)
+    Process: 1026 ExecStartPre=/usr/sbin/nginx -t -q -g daemon on; master_process on; (code=exited, status=0/SUCCESS)
+    Process: 1106 ExecStart=/usr/sbin/nginx -g daemon on; master_process on; (code=exited, status=0/SUCCESS)
+   Main PID: 1107 (nginx)
+      Tasks: 5 (limit: 6929)
+     Memory: 5.7M
+     CGroup: /system.slice/nginx.service
+             ├─1107 nginx: master process /usr/sbin/nginx -g daemon on; master_process on;
+             ├─1108 nginx: worker process
+             ├─1109 nginx: worker process
+             ├─1110 nginx: worker process
+             └─1111 nginx: worker process
+```
+
+Next you need to install certbot which will handle all of the certificate automation for your web server and domain. To install certbot it is required to have `snapd` on Ubuntu. If you haven't installed it yet, please review the instructions on [snapcraft's site for installation](https://snapcraft.io/docs/installing-snapd/).
+
+Once you are good to go, ensure your snapd version is up to date:
+
+``` bash
+> sudo snap install core; sudo snap refresh core
+```
+
+If you have any Certbot packages installed via your package manager, you should remove them before installing Certbot:
+
+``` bash
+> sudo apt remove certbot
+```
+
+Finally, install Certbot and prepare the Certbot command.
+
+``` bash
+> sudo snap install --classic certbot
+> sudo ln -s /snap/bin/certbot /usr/bin/certbot
+```
+
+**Step 2: Adjust the firewall**
+
+The next step is to allow HTTP and HTTPS traffic through your firewall. This is different for each depending on which firewall you use, and how you have it configured. In this example, we are using `ufw`.
+
+We'll start by enabling `ufw` if it isn't already:
+
+``` bash
+> sudo ufw enable
+```
+
+You can list the application configurations that ufw knows how to work with
+
+``` bash
[…]
