### 2022-11-21 0cb0802
Bump numpy from 1.23.4 to 1.23.5
Bumps [numpy](https://github.com/numpy/numpy) from 1.23.4 to 1.23.5.
- [Release notes](https://github.com/numpy/numpy/releases)
- [Changelog](https://github.com/numpy/numpy/blob/main/doc/RELEASE_WALKTHROUGH.rst)
- [Commits](https://github.com/numpy/numpy/compare/v1.23.4...v1.23.5)

---
updated-dependencies:
- dependency-name: numpy
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
