### 2022-11-21 3f9dacc
Bump ccxt from 2.1.75 to 2.1.96
Bumps [ccxt](https://github.com/ccxt/ccxt) from 2.1.75 to 2.1.96.
- [Release notes](https://github.com/ccxt/ccxt/releases)
- [Changelog](https://github.com/ccxt/ccxt/blob/master/exchanges.cfg)
- [Commits](https://github.com/ccxt/ccxt/compare/2.1.75...2.1.96)

---
updated-dependencies:
- dependency-name: ccxt
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
