### 2022-11-21 0fa5217
Improve protection setup
lock_pair should be called when the order closes, not when the exit order is placed.
it should also be called for stoploss orders, too.

closes #7783
