### 2022-11-21 d9d7df7
fix tests, log unknown errors
