### 2022-11-21 8cb2b46
Improve proxy docs
closes #7769
