### 2022-11-22 4d6c52b
add "how to run tests"
