### 2022-11-22 fff745f
add map to nginx config
