### 2022-11-22 c01f25d
update code to freqai_backtest_live_models only from historic predictions
