+++ freqtrade/freqai/data_kitchen.py
@@ -1,7 +1,7 @@
 import copy
 import logging
 import shutil
-from datetime import datetime, timedelta, timezone
+from datetime import datetime, timezone
 from math import cos, sin
 from pathlib import Path
 from typing import Any, Dict, List, Tuple
@@ -86,14 +86,7 @@ class FreqaiDataKitchen:
         if not self.live:
             self.full_path = self.get_full_models_path(self.config)
 
-            if self.backtest_live_models:
-                if self.pair and not (
-                    self.freqai_config.get("backtest_using_historic_predictions", True)
-                ):
-                    self.set_timerange_from_ready_models()
-                    (self.training_timeranges,
-                     self.backtesting_timeranges) = self.split_timerange_live_models()
-            else:
+            if not self.backtest_live_models:
                 self.full_timerange = self.create_fulltimerange(
                     self.config["timerange"], self.freqai_config.get("train_period_days", 0)
                 )
@@ -458,28 +451,28 @@ class FreqaiDataKitchen:
         # print(tr_training_list, tr_backtesting_list)
         return tr_training_list_timerange, tr_backtesting_list_timerange
 
-    def split_timerange_live_models(
-        self
-    ) -> Tuple[list, list]:
+    # def split_timerange_live_models(
+    #     self
+    # ) -> Tuple[list, list]:
 
-        tr_backtesting_list_timerange = []
-        asset = self.pair.split("/")[0]
-        if asset not in self.backtest_live_models_data["assets_end_dates"]:
-            raise OperationalException(
-                f"Model not available for pair {self.pair}. "
-                "Please, try again after removing this pair from the configuration file."
-            )
-        asset_data = self.backtest_live_models_data["assets_end_dates"][asset]
-        backtesting_timerange = self.backtest_live_models_data["backtesting_timerange"]
-        model_end_dates = [x for x in asset_data]
-        model_end_dates.append(backtesting_timerange.stopts)
-        model_end_dates.sort()
-        for index, item in enumerate(model_end_dates):
-            if len(model_end_dates) > (index + 1):
-                tr_to_add = TimeRange("date", "date", item, model_end_dates[index + 1])
-                tr_backtesting_list_timerange.append(tr_to_add)
+    #     tr_backtesting_list_timerange = []
+    #     asset = self.pair.split("/")[0]
+    #     if asset not in self.backtest_live_models_data["assets_end_dates"]:
+    #         raise OperationalException(
+    #             f"Model not available for pair {self.pair}. "
+    #             "Please, try again after removing this pair from the configuration file."
+    #         )
+    #     asset_data = self.backtest_live_models_data["assets_end_dates"][asset]
+    #     backtesting_timerange = self.backtest_live_models_data["backtesting_timerange"]
+    #     model_end_dates = [x for x in asset_data]
+    #     model_end_dates.append(backtesting_timerange.stopts)
+    #     model_end_dates.sort()
+    #     for index, item in enumerate(model_end_dates):
+    #         if len(model_end_dates) > (index + 1):
+    #             tr_to_add = TimeRange("date", "date", item, model_end_dates[index + 1])
+    #             tr_backtesting_list_timerange.append(tr_to_add)
 
-        return tr_backtesting_list_timerange, tr_backtesting_list_timerange
+    #     return tr_backtesting_list_timerange, tr_backtesting_list_timerange
 
     def slice_dataframe(self, timerange: TimeRange, df: DataFrame) -> DataFrame:
         """
@@ -1371,17 +1364,6 @@ class FreqaiDataKitchen:
             )
             return False
 
-    def set_timerange_from_ready_models(self):
-        backtesting_timerange, \
-            assets_end_dates = (
-                self.get_timerange_and_assets_end_dates_from_ready_models(self.full_path))
-
-        self.backtest_live_models_data = {
[…]
