### 2022-11-22 a544277
ensure only broadcasting to subscribed topics
