### 2022-11-22 bd05f85
Simplify ssl documentation
