### 2022-11-23 e5fc21f
Fix broken table rendering
