### 2022-11-24 8f1a8c7
Add freqairl docker build process
