### 2022-11-24 44b042b
remove unused function
