### 2022-11-24 8855e36
reduce freqai testing time by reducing retrain frequency and number of features
