+++ tests/freqai/test_models/ReinforcementLearner_test_4ac.py
@@ -1,56 +1,18 @@
 import logging
-from pathlib import Path
-from typing import Any, Dict
 
 import numpy as np
-import torch as th
 
-from freqtrade.freqai.data_kitchen import FreqaiDataKitchen
 from freqtrade.freqai.RL.Base4ActionRLEnv import Actions, Base4ActionRLEnv, Positions
-from freqtrade.freqai.RL.BaseReinforcementLearningModel import BaseReinforcementLearningModel
+from freqtrade.freqai.prediction_models.ReinforcementLearner import ReinforcementLearner
 
 
 logger = logging.getLogger(__name__)
 
 
-class ReinforcementLearner_test_4ac(BaseReinforcementLearningModel):
+class ReinforcementLearner_test_4ac(ReinforcementLearner):
     """
     User created Reinforcement Learning Model prediction model.
     """
-
-    def fit(self, data_dictionary: Dict[str, Any], dk: FreqaiDataKitchen, **kwargs):
-
-        train_df = data_dictionary["train_features"]
-        total_timesteps = self.freqai_info["rl_config"]["train_cycles"] * len(train_df)
-
-        policy_kwargs = dict(activation_fn=th.nn.ReLU,
-                             net_arch=[64, 64])
-
-        if dk.pair not in self.dd.model_dictionary or not self.continual_learning:
-            model = self.MODELCLASS(self.policy_type, self.train_env, policy_kwargs=policy_kwargs,
-                                    tensorboard_log=Path(
-                                        dk.full_path / "tensorboard" / dk.pair.split('/')[0]),
-                                    **self.freqai_info['model_training_parameters']
-                                    )
-        else:
-            logger.info('Continual training activated - starting training from previously '
-                        'trained agent.')
-            model = self.dd.model_dictionary[dk.pair]
-            model.set_env(self.train_env)
-
-        model.learn(
-            total_timesteps=int(total_timesteps),
-            callback=self.eval_callback
-        )
-
-        if Path(dk.data_path / "best_model.zip").is_file():
-            logger.info('Callback found a best model.')
-            best_model = self.MODELCLASS.load(dk.data_path / "best_model")
-            return best_model
-
-        logger.info('Couldnt find best model, using final model instead.')
-
-        return model
 
     class MyRLEnv(Base4ActionRLEnv):
         """
