+++ tests/freqai/test_models/ReinforcementLearner_test_4ac.py
@@ -2,8 +2,8 @@ import logging
 
 import numpy as np
 
-from freqtrade.freqai.RL.Base4ActionRLEnv import Actions, Base4ActionRLEnv, Positions
 from freqtrade.freqai.prediction_models.ReinforcementLearner import ReinforcementLearner
+from freqtrade.freqai.RL.Base4ActionRLEnv import Actions, Base4ActionRLEnv, Positions
 
 
 logger = logging.getLogger(__name__)
