### 2022-11-24 2e82e67
move data_provider cleanup to shutdown()
