### 2022-11-24 fc59b02
prevent ws endpoint from running without valid token
