### 2022-11-25 7b0a76f
Improve typehint
