### 2022-11-25 5e6cda1
Update method name for trade fee updating
