### 2022-11-25 c593cdc
Improve type hints
