### 2022-11-25 1b3e62b
Lock execute_entry to prevent timing hickups
