### 2022-11-25 0f97ef0
Reset stoploss_order_id when order is canceled
closes #7766
