### 2022-11-25 8ee8b6e
Improve hyperopt list output
closes  #7789
