### 2022-11-25 8c014bd
Export trade-counts to csv
closes #7789
