### 2022-11-25 3918172
Tidy up complex functions
