### 2022-11-25 afc00bc
log warning if channel too far behind, add docstrings to message stream
