### 2022-11-25 f268187
offload initial df computation to thread
