### 2022-11-26 9660e44
use new channel apis in emc, extend analyzed df to include list of dates for candles
