### 2022-11-26 4cbb334
change how missing candles will be handled
