+++ docs/JOSS_paper/paper.md
@@ -201,10 +201,11 @@ However, `FreqAI` also benefits from `Freqtrade` docker distributions, and can b
 
 We would like to acknowledge various beta testers of `FreqAI`:
 
-- Richárd Józsa
-- Juha Nykänen
-- Emre Suzen
-- Salah Lamkadem
+- Longlong Yu (@lolongcovas)
+- Richárd Józsa (@richardjozsa)
+- Juha Nykänen (@suikula)
+- Emre Suzen (@aemr3)
+- Salah Lamkadem (@ikonx)
 
 As well as various `Freqtrade` [developers](https://github.com/freqtrade/freqtrade/graphs/contributors) maintaining tangential, yet essential, modules.
 
