### 2022-11-26 9f13d99
improve parameter table, add better documentation for custom calculate_reward, add various helpful notes in docstrings etc
