### 2022-11-26 81fd2e5
ensure typing, remove unsued code
