### 2022-12-07 #4864 @Fei-Tao
Hi @kyleniemeyer, thanks for the clarification. Hi @openjournals/dsais-eics, we got an issue to obtain the comments from the last reviewer. But we have received responses from two reviewers. Can we move forward to accept this submission? Thank you for your time.
