### 2022-12-07 #4864 @gkthiruvathukal
@Fei-Tao Yes, please proceed to remove the reviewer as suggested by @arfon.  Then you can make your recommendation as editor so I can proceed with the final checks.
