### 2022-12-10 #4864 @robcaulk
Thanks @Fei-Tao, the DOI is fixed [here](https://github.com/freqtrade/freqtrade/actions/runs/3640446380)
