### 2022-12-11 #4864 @robcaulk
@Fei-Tao sure, here is the release:

https://github.com/freqtrade/freqtrade/releases/tag/JOSS_v1

And here is the published archive on Zenodo:

https://zenodo.org/record/7424314#.Y5X6xTPMJkg

Please let me know if there is anything else that we can do to help finalize the publication process.
