### 2022-12-12 #4864 @robcaulk
Hey @Fei-Tao ,

I have fixed the github usernames in the acknowledgements so that they are no longer mistaken as references.

Here is the new archive version:

https://zenodo.org/record/7429524#.Y5dlNjPMJkg

And here is the new release:

https://github.com/freqtrade/freqtrade/releases/tag/JOSS_v2

Please let me know if there are any other changes we should make,

Robert
