### 2022-12-12 #4864 @robcaulk
@Fei-Tao we noticed that the reviewer list still includes annielytix, should we remove her name from this list?

Also, we noticed the footer has “Ph.D et Al” which should say “Caulk et al.” Is this something we can edit before acceptance? If so, should we update the archive to reflect these changes?
