### 2022-12-12 #4864 @Fei-Tao
@robcaulk the command worked this time. The additional reviwer has been removed successfully. Please go head to modify the footnote.
