### 2022-12-13 #4864 @robcaulk
@Fei-Tao great, here is the updated archive and release:

https://zenodo.org/record/7431513#.Y5gy3zPMJkg

https://github.com/freqtrade/freqtrade/releases/tag/JOSS_v3

Thanks for your patience :-)
