### 2022-12-15 #4864 @robcaulk
@Fei-Tao @arfon Thanks, this looks great. The paper is perfect. I noticed the correct tags are shown on the website to the right of the paper, but it seems the tags on the top (see image below) are relatively random. Can we change these top tags to match the the tags shown to the right of the document?

![image](https://user-images.githubusercontent.com/20807672/207971622-6dc9ef7a-d0b1-4897-a647-cb035c6f5cee.png)
