### 2022-12-18 freqtrade/freqtrade#7668 @xmatthias
Given that the JOSS paper has now been accepted, i assume we no longer need this issue.
