### 2022-12-26 #4864 @arfon
> Can we change these top tags to match the the tags shown to the right of the document?

These tags are generated by the Linguist language classification library which is the same one used here by GitHub. I'll remove everything but Python to make it less confusing.
